Move some APIs to frontend to admin * Move GetReplicationMessagesResponse / GetDomainReplicationMessagesResponse / ReapplyEvents to admin endpoint
