+++ .gen/go/admin/adminserviceserver/server.go
@@ -28,6 +28,7 @@ package adminserviceserver
 import (
 	context "context"
 	admin "github.com/uber/cadence/.gen/go/admin"
+	replicator "github.com/uber/cadence/.gen/go/replicator"
 	shared "github.com/uber/cadence/.gen/go/shared"
 	wire "go.uber.org/thriftrw/wire"
 	transport "go.uber.org/yarpc/api/transport"
@@ -60,6 +61,16 @@ type Interface interface {
 		Request *admin.DescribeWorkflowExecutionRequest,
 	) (*admin.DescribeWorkflowExecutionResponse, error)
 
+	GetDomainReplicationMessages(
+		ctx context.Context,
+		Request *replicator.GetDomainReplicationMessagesRequest,
+	) (*replicator.GetDomainReplicationMessagesResponse, error)
+
+	GetReplicationMessages(
+		ctx context.Context,
+		Request *replicator.GetReplicationMessagesRequest,
+	) (*replicator.GetReplicationMessagesResponse, error)
+
 	GetWorkflowExecutionRawHistory(
 		ctx context.Context,
 		GetRequest *admin.GetWorkflowExecutionRawHistoryRequest,
@@ -69,6 +80,11 @@ type Interface interface {
 		ctx context.Context,
 		GetRequest *admin.GetWorkflowExecutionRawHistoryV2Request,
 	) (*admin.GetWorkflowExecutionRawHistoryV2Response, error)
+
+	ReapplyEvents(
+		ctx context.Context,
+		ReapplyEventsRequest *shared.ReapplyEventsRequest,
+	) error
 
 	RemoveTask(
 		ctx context.Context,
@@ -143,6 +159,28 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 			},
 
 			thrift.Method{
+				Name: "GetDomainReplicationMessages",
+				HandlerSpec: thrift.HandlerSpec{
+
+					Type:  transport.Unary,
+					Unary: thrift.UnaryHandler(h.GetDomainReplicationMessages),
+				},
+				Signature:    "GetDomainReplicationMessages(Request *replicator.GetDomainReplicationMessagesRequest) (*replicator.GetDomainReplicationMessagesResponse)",
+				ThriftModule: admin.ThriftModule,
+			},
+
+			thrift.Method{
+				Name: "GetReplicationMessages",
+				HandlerSpec: thrift.HandlerSpec{
+
+					Type:  transport.Unary,
+					Unary: thrift.UnaryHandler(h.GetReplicationMessages),
+				},
+				Signature:    "GetReplicationMessages(Request *replicator.GetReplicationMessagesRequest) (*replicator.GetReplicationMessagesResponse)",
+				ThriftModule: admin.ThriftModule,
+			},
+
+			thrift.Method{
 				Name: "GetWorkflowExecutionRawHistory",
 				HandlerSpec: thrift.HandlerSpec{
 
@@ -165,6 +203,17 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 			},
 
 			thrift.Method{
+				Name: "ReapplyEvents",
+				HandlerSpec: thrift.HandlerSpec{
+
+					Type:  transport.Unary,
+					Unary: thrift.UnaryHandler(h.ReapplyEvents),
+				},
+				Signature:    "ReapplyEvents(ReapplyEventsRequest *shared.ReapplyEventsRequest)",
+				ThriftModule: admin.ThriftModule,
+			},
+
+			thrift.Method{
 				Name: "RemoveTask",
 				HandlerSpec: thrift.HandlerSpec{
 
@@ -177,7 +226,7 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 		},
 	}
 
-	procedures := make([]transport.Procedure, 0, 8)
+	procedures := make([]transport.Procedure, 0, 11)
 	procedures = append(procedures, thrift.BuildProcedures(service, opts...)...)
 	return procedures
 }
@@ -279,6 +328,44 @@ func (h handler) DescribeWorkflowExecution(ctx context.Context, body wire.Value)
 	return response, err
 }
 
+func (h handler) GetDomainReplicationMessages(ctx context.Context, body wire.Value) (thrift.Response, error) {
+	var args admin.AdminService_GetDomainReplicationMessages_Args
+	if err := args.FromWire(body); err != nil {
+		return thrift.Response{}, err
+	}
+
+	success, err := h.impl.GetDomainReplicationMessages(ctx, args.Request)
+
+	hadError := err != nil
+	result, err := admin.AdminService_GetDomainReplicationMessages_Helper.WrapResponse(success, err)
+
+	var response thrift.Response
+	if err == nil {
+		response.IsApplicationError = hadError
+		response.Body = result
+	}
+	return response, err
+}
+
+func (h handler) GetReplicationMessages(ctx context.Context, body wire.Value) (thrift.Response, error) {
+	var args admin.AdminService_GetReplicationMessages_Args
+	if err := args.FromWire(body); err != nil {
+		return thrift.Response{}, err
+	}
+
+	success, err := h.impl.GetReplicationMessages(ctx, args.Request)
+
+	hadError := err != nil
+	result, err := admin.AdminService_GetReplicationMessages_Helper.WrapResponse(success, err)
+
+	var response thrift.Response
+	if err == nil {
+		response.IsApplicationError = hadError
+		response.Body = result
+	}
+	return response, err
+}
+
 func (h handler) GetWorkflowExecutionRawHistory(ctx context.Context, body wire.Value) (thrift.Response, error) {
 	var args admin.AdminService_GetWorkflowExecutionRawHistory_Args
 	if err := args.FromWire(body); err != nil {
@@ -308,6 +395,25 @@ func (h handler) GetWorkflowExecutionRawHistoryV2(ctx context.Context, body wire
 
 	hadError := err != nil
 	result, err := admin.AdminService_GetWorkflowExecutionRawHistoryV2_Helper.WrapResponse(success, err)
+
+	var response thrift.Response
+	if err == nil {
+		response.IsApplicationError = hadError
+		response.Body = result
+	}
+	return response, err
+}
+
+func (h handler) ReapplyEvents(ctx context.Context, body wire.Value) (thrift.Response, error) {
+	var args admin.AdminService_ReapplyEvents_Args
+	if err := args.FromWire(body); err != nil {
+		return thrift.Response{}, err
+	}
+
+	err := h.impl.ReapplyEvents(ctx, args.ReapplyEventsRequest)
+
+	hadError := err != nil
+	result, err := admin.AdminService_ReapplyEvents_Helper.WrapResponse(err)
 
 	var response thrift.Response
 	if err == nil {
