+++ .gen/go/cadence/cadence.go
@@ -42,7 +42,7 @@ var ThriftModule = &thriftreflect.ThriftModule{
 	Name:     "cadence",
 	Package:  "github.com/uber/cadence/.gen/go/cadence",
 	FilePath: "cadence.thrift",
-	SHA1:     "dde2814954c9f66f7bbd85467eecd7495f10e9bf",
+	SHA1:     "e58a3fd63f822cc28c5197d360d5e636e7f4dfd4",
 	Includes: []*thriftreflect.ThriftModule{
 		replicator.ThriftModule,
 		shared.ThriftModule,
@@ -50,7 +50,7 @@ var ThriftModule = &thriftreflect.ThriftModule{
 	Raw: rawIDL,
 }
 
-const rawIDL = "// Copyright (c) 2017 Uber Technologies, Inc.\n//\n// Permission is hereby granted, free of charge, to any person obtaining a copy\n// of this software and associated documentation files (the \"Software\"), to deal\n// in the Software without restriction, including without limitation the rights\n// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell\n// copies of the Software, and to permit persons to whom the Software is\n// furnished to do so, subject to the following conditions:\n//\n// The above copyright notice and this permission notice shall be included in\n// all copies or substantial portions of the Software.\n//\n// THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,\n// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER\n// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,\n// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN\n// THE SOFTWARE.\n\ninclude \"shared.thrift\"\ninclude \"replicator.thrift\"\n\nnamespace java com.uber.cadence\n\n/**\n* WorkflowService API is exposed to provide support for long running applications.  Application is expected to call\n* StartWorkflowExecution to create an instance for each instance of long running workflow.  Such applications are expected\n* to have a worker which regularly polls for DecisionTask and ActivityTask from the WorkflowService.  For each\n* DecisionTask, application is expected to process the history of events for that session and respond back with next\n* decisions.  For each ActivityTask, application is expected to execute the actual logic for that task and respond back\n* with completion or failure.  Worker is expected to regularly heartbeat while activity task is running.\n**/\nservice WorkflowService {\n  /**\n  * RegisterDomain creates a new domain which can be used as a container for all resources.  Domain is a top level\n  * entity within Cadence, used as a container for all resources like workflow executions, tasklists, etc.  Domain\n  * acts as a sandbox and provides isolation for all resources within the domain.  All resources belongs to exactly one\n  * domain.\n  **/\n  void RegisterDomain(1: shared.RegisterDomainRequest registerRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.DomainAlreadyExistsError domainExistsError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * DescribeDomain returns the information and configuration for a registered domain.\n  **/\n  shared.DescribeDomainResponse DescribeDomain(1: shared.DescribeDomainRequest describeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n    * ListDomains returns the information and configuration for all domains.\n    **/\n    shared.ListDomainsResponse ListDomains(1: shared.ListDomainsRequest listRequest)\n      throws (\n        1: shared.BadRequestError badRequestError,\n        3: shared.EntityNotExistsError entityNotExistError,\n        4: shared.ServiceBusyError serviceBusyError,\n        5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n      )\n\n  /**\n  * UpdateDomain is used to update the information and configuration for a registered domain.\n  **/\n  shared.UpdateDomainResponse UpdateDomain(1: shared.UpdateDomainRequest updateRequest)\n      throws (\n        1: shared.BadRequestError badRequestError,\n        3: shared.EntityNotExistsError entityNotExistError,\n        4: shared.ServiceBusyError serviceBusyError,\n        5: shared.DomainNotActiveError domainNotActiveError,\n        6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n      )\n\n  /**\n  * DeprecateDomain us used to update status of a registered domain to DEPRECATED.  Once the domain is deprecated\n  * it cannot be used to start new workflow executions.  Existing workflow executions will continue to run on\n  * deprecated domains.\n  **/\n  void DeprecateDomain(1: shared.DeprecateDomainRequest deprecateRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * StartWorkflowExecution starts a new long running workflow instance.  It will create the instance with\n  * 'WorkflowExecutionStarted' event in history and also schedule the first DecisionTask for the worker to make the\n  * first decision for this instance.  It will return 'WorkflowExecutionAlreadyStartedError', if an instance already\n  * exists with same workflowId.\n  **/\n  shared.StartWorkflowExecutionResponse StartWorkflowExecution(1: shared.StartWorkflowExecutionRequest startRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.WorkflowExecutionAlreadyStartedError sessionAlreadyExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.EntityNotExistsError entityNotExistError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * Returns the history of specified workflow execution.  It fails with 'EntityNotExistError' if speficied workflow\n  * execution in unknown to the service.\n  **/\n  shared.GetWorkflowExecutionHistoryResponse GetWorkflowExecutionHistory(1: shared.GetWorkflowExecutionHistoryRequest getRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * PollForDecisionTask is called by application worker to process DecisionTask from a specific taskList.  A\n  * DecisionTask is dispatched to callers for active workflow executions, with pending decisions.\n  * Application is then expected to call 'RespondDecisionTaskCompleted' API when it is done processing the DecisionTask.\n  * It will also create a 'DecisionTaskStarted' event in the history for that session before handing off DecisionTask to\n  * application worker.\n  **/\n  shared.PollForDecisionTaskResponse PollForDecisionTask(1: shared.PollForDecisionTaskRequest pollRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.ServiceBusyError serviceBusyError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.EntityNotExistsError entityNotExistError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondDecisionTaskCompleted is called by application worker to complete a DecisionTask handed as a result of\n  * 'PollForDecisionTask' API call.  Completing a DecisionTask will result in new events for the workflow execution and\n  * potentially new ActivityTask being created for corresponding decisions.  It will also create a DecisionTaskCompleted\n  * event in the history for that session.  Use the 'taskToken' provided as response of PollForDecisionTask API call\n  * for completing the DecisionTask.\n  * The response could contain a new decision task if there is one or if the request asking for one.\n  **/\n  shared.RespondDecisionTaskCompletedResponse RespondDecisionTaskCompleted(1: shared.RespondDecisionTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondDecisionTaskFailed is called by application worker to indicate failure.  This results in\n  * DecisionTaskFailedEvent written to the history and a new DecisionTask created.  This API can be used by client to\n  * either clear sticky tasklist or report any panics during DecisionTask processing.  Cadence will only append first\n  * DecisionTaskFailed event to the history of workflow execution for consecutive failures.\n  **/\n  void RespondDecisionTaskFailed(1: shared.RespondDecisionTaskFailedRequest failedRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * PollForActivityTask is called by application worker to process ActivityTask from a specific taskList.  ActivityTask\n  * is dispatched to callers whenever a ScheduleTask decision is made for a workflow execution.\n  * Application is expected to call 'RespondActivityTaskCompleted' or 'RespondActivityTaskFailed' once it is done\n  * processing the task.\n  * Application also needs to call 'RecordActivityTaskHeartbeat' API within 'heartbeatTimeoutSeconds' interval to\n  * prevent the task from getting timed out.  An event 'ActivityTaskStarted' event is also written to workflow execution\n  * history before the ActivityTask is dispatched to application worker.\n  **/\n  shared.PollForActivityTaskResponse PollForActivityTask(1: shared.PollForActivityTaskRequest pollRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.ServiceBusyError serviceBusyError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.EntityNotExistsError entityNotExistError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RecordActivityTaskHeartbeat is called by application worker while it is processing an ActivityTask.  If worker fails\n  * to heartbeat within 'heartbeatTimeoutSeconds' interval for the ActivityTask, then it will be marked as timedout and\n  * 'ActivityTaskTimedOut' event will be written to the workflow history.  Calling 'RecordActivityTaskHeartbeat' will\n  * fail with 'EntityNotExistsError' in such situations.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for heartbeating.\n  **/\n  shared.RecordActivityTaskHeartbeatResponse RecordActivityTaskHeartbeat(1: shared.RecordActivityTaskHeartbeatRequest heartbeatRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RecordActivityTaskHeartbeatByID is called by application worker while it is processing an ActivityTask.  If worker fails\n  * to heartbeat within 'heartbeatTimeoutSeconds' interval for the ActivityTask, then it will be marked as timedout and\n  * 'ActivityTaskTimedOut' event will be written to the workflow history.  Calling 'RecordActivityTaskHeartbeatByID' will\n  * fail with 'EntityNotExistsError' in such situations.  Instead of using 'taskToken' like in RecordActivityTaskHeartbeat,\n  * use Domain, WorkflowID and ActivityID\n  **/\n  shared.RecordActivityTaskHeartbeatResponse RecordActivityTaskHeartbeatByID(1: shared.RecordActivityTaskHeartbeatByIDRequest heartbeatRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCompleted is called by application worker when it is done processing an ActivityTask.  It will\n  * result in a new 'ActivityTaskCompleted' event being written to the workflow history and a new DecisionTask\n  * created for the workflow so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskCompleted(1: shared.RespondActivityTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCompletedByID is called by application worker when it is done processing an ActivityTask.\n  * It will result in a new 'ActivityTaskCompleted' event being written to the workflow history and a new DecisionTask\n  * created for the workflow so new decisions could be made.  Similar to RespondActivityTaskCompleted but use Domain,\n  * WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskCompletedByID(1: shared.RespondActivityTaskCompletedByIDRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskFailed is called by application worker when it is done processing an ActivityTask.  It will\n  * result in a new 'ActivityTaskFailed' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskFailed(1: shared.RespondActivityTaskFailedRequest failRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskFailedByID is called by application worker when it is done processing an ActivityTask.\n  * It will result in a new 'ActivityTaskFailed' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Similar to RespondActivityTaskFailed but use\n  * Domain, WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskFailedByID(1: shared.RespondActivityTaskFailedByIDRequest failRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCanceled is called by application worker when it is successfully canceled an ActivityTask.  It will\n  * result in a new 'ActivityTaskCanceled' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void RespondActivityTaskCanceled(1: shared.RespondActivityTaskCanceledRequest canceledRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCanceledByID is called by application worker when it is successfully canceled an ActivityTask.\n  * It will result in a new 'ActivityTaskCanceled' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Similar to RespondActivityTaskCanceled but use\n  * Domain, WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void RespondActivityTaskCanceledByID(1: shared.RespondActivityTaskCanceledByIDRequest canceledRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RequestCancelWorkflowExecution is called by application worker when it wants to request cancellation of a workflow instance.\n  * It will result in a new 'WorkflowExecutionCancelRequested' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made. It fails with 'EntityNotExistsError' if the workflow is not valid\n  * anymore due to completion or doesn't exist.\n  **/\n  void RequestCancelWorkflowExecution(1: shared.RequestCancelWorkflowExecutionRequest cancelRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.CancellationAlreadyRequestedError cancellationAlreadyRequestedError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.LimitExceededError limitExceededError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * SignalWorkflowExecution is used to send a signal event to running workflow execution.  This results in\n  * WorkflowExecutionSignaled event recorded in the history and a decision task being created for the execution.\n  **/\n  void SignalWorkflowExecution(1: shared.SignalWorkflowExecutionRequest signalRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * SignalWithStartWorkflowExecution is used to ensure sending signal to a workflow.\n  * If the workflow is running, this results in WorkflowExecutionSignaled event being recorded in the history\n  * and a decision task being created for the execution.\n  * If the workflow is not running or not found, this results in WorkflowExecutionStarted and WorkflowExecutionSignaled\n  * events being recorded in history, and a decision task being created for the execution\n  **/\n  shared.StartWorkflowExecutionResponse SignalWithStartWorkflowExecution(1: shared.SignalWithStartWorkflowExecutionRequest signalWithStartRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.WorkflowExecutionAlreadyStartedError workflowAlreadyStartedError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n    * ResetWorkflowExecution reset an existing workflow execution to DecisionTaskCompleted event(exclusive).\n    * And it will immediately terminating the current execution instance.\n    **/\n  shared.ResetWorkflowExecutionResponse ResetWorkflowExecution(1: shared.ResetWorkflowExecutionRequest resetRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * TerminateWorkflowExecution terminates an existing workflow execution by recording WorkflowExecutionTerminated event\n  * in the history and immediately terminating the execution instance.\n  **/\n  void TerminateWorkflowExecution(1: shared.TerminateWorkflowExecutionRequest terminateRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListOpenWorkflowExecutions is a visibility API to list the open executions in a specific domain.\n  **/\n  shared.ListOpenWorkflowExecutionsResponse ListOpenWorkflowExecutions(1: shared.ListOpenWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListClosedWorkflowExecutions is a visibility API to list the closed executions in a specific domain.\n  **/\n  shared.ListClosedWorkflowExecutionsResponse ListClosedWorkflowExecutions(1: shared.ListClosedWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListWorkflowExecutions is a visibility API to list workflow executions in a specific domain.\n  **/\n  shared.ListWorkflowExecutionsResponse ListWorkflowExecutions(1: shared.ListWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListArchivedWorkflowExecutions is a visibility API to list archived workflow executions in a specific domain.\n  **/\n  shared.ListArchivedWorkflowExecutionsResponse ListArchivedWorkflowExecutions(1: shared.ListArchivedWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ScanWorkflowExecutions is a visibility API to list large amount of workflow executions in a specific domain without order.\n  **/\n  shared.ListWorkflowExecutionsResponse ScanWorkflowExecutions(1: shared.ListWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * CountWorkflowExecutions is a visibility API to count of workflow executions in a specific domain.\n  **/\n  shared.CountWorkflowExecutionsResponse CountWorkflowExecutions(1: shared.CountWorkflowExecutionsRequest countRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * GetSearchAttributes is a visibility API to get all legal keys that could be used in list APIs\n  **/\n  shared.GetSearchAttributesResponse GetSearchAttributes()\n    throws (\n      2: shared.ServiceBusyError serviceBusyError,\n      3: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondQueryTaskCompleted is called by application worker to complete a QueryTask (which is a DecisionTask for query)\n  * as a result of 'PollForDecisionTask' API call. Completing a QueryTask will unblock the client call to 'QueryWorkflow'\n  * API and return the query result to client as a response to 'QueryWorkflow' API call.\n  **/\n  void RespondQueryTaskCompleted(1: shared.RespondQueryTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * Reset the sticky tasklist related information in mutable state of a given workflow.\n  * Things cleared are:\n  * 1. StickyTaskList\n  * 2. StickyScheduleToStartTimeout\n  * 3. ClientLibraryVersion\n  * 4. ClientFeatureVersion\n  * 5. ClientImpl\n  **/\n  shared.ResetStickyTaskListResponse ResetStickyTaskList(1: shared.ResetStickyTaskListRequest resetRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * QueryWorkflow returns query result for a specified workflow execution\n  **/\n  shared.QueryWorkflowResponse QueryWorkflow(1: shared.QueryWorkflowRequest queryRequest)\n\tthrows (\n\t  1: shared.BadRequestError badRequestError,\n\t  3: shared.EntityNotExistsError entityNotExistError,\n\t  4: shared.QueryFailedError queryFailedError,\n\t  5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n\t)\n\n  /**\n  * DescribeWorkflowExecution returns information about the specified workflow execution.\n  **/\n  shared.DescribeWorkflowExecutionResponse DescribeWorkflowExecution(1: shared.DescribeWorkflowExecutionRequest describeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * DescribeTaskList returns information about the target tasklist, right now this API returns the\n  * pollers which polled this tasklist in last few minutes.\n  **/\n  shared.DescribeTaskListResponse DescribeTaskList(1: shared.DescribeTaskListRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  replicator.GetReplicationMessagesResponse GetReplicationMessages(1: replicator.GetReplicationMessagesRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.LimitExceededError limitExceededError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  replicator.GetDomainReplicationMessagesResponse GetDomainReplicationMessages(1: replicator.GetDomainReplicationMessagesRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.LimitExceededError limitExceededError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ReapplyEvents applies stale events to the current workflow and current run\n  **/\n  void ReapplyEvents(1: shared.ReapplyEventsRequest reapplyEventsRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.DomainNotActiveError domainNotActiveError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.EntityNotExistsError entityNotExistError,\n    )\n\n  /**\n  * GetClusterInfo returns information about cadence cluster\n  **/\n  shared.ClusterInfo GetClusterInfo()\n    throws (\n      1: shared.InternalServiceError internalServiceError,\n      2: shared.ServiceBusyError serviceBusyError,\n    )\n\n   /**\n   * ReapplyEvents applies stale events to the current workflow and current run\n   **/\n  shared.ListTaskListPartitionsResponse ListTaskListPartitions(1: shared.ListTaskListPartitionsRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n    )\n}\n"
+const rawIDL = "// Copyright (c) 2017 Uber Technologies, Inc.\n//\n// Permission is hereby granted, free of charge, to any person obtaining a copy\n// of this software and associated documentation files (the \"Software\"), to deal\n// in the Software without restriction, including without limitation the rights\n// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell\n// copies of the Software, and to permit persons to whom the Software is\n// furnished to do so, subject to the following conditions:\n//\n// The above copyright notice and this permission notice shall be included in\n// all copies or substantial portions of the Software.\n//\n// THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,\n// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER\n// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,\n// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN\n// THE SOFTWARE.\n\ninclude \"shared.thrift\"\ninclude \"replicator.thrift\"\n\nnamespace java com.uber.cadence\n\n/**\n* WorkflowService API is exposed to provide support for long running applications.  Application is expected to call\n* StartWorkflowExecution to create an instance for each instance of long running workflow.  Such applications are expected\n* to have a worker which regularly polls for DecisionTask and ActivityTask from the WorkflowService.  For each\n* DecisionTask, application is expected to process the history of events for that session and respond back with next\n* decisions.  For each ActivityTask, application is expected to execute the actual logic for that task and respond back\n* with completion or failure.  Worker is expected to regularly heartbeat while activity task is running.\n**/\nservice WorkflowService {\n  /**\n  * RegisterDomain creates a new domain which can be used as a container for all resources.  Domain is a top level\n  * entity within Cadence, used as a container for all resources like workflow executions, tasklists, etc.  Domain\n  * acts as a sandbox and provides isolation for all resources within the domain.  All resources belongs to exactly one\n  * domain.\n  **/\n  void RegisterDomain(1: shared.RegisterDomainRequest registerRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.DomainAlreadyExistsError domainExistsError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * DescribeDomain returns the information and configuration for a registered domain.\n  **/\n  shared.DescribeDomainResponse DescribeDomain(1: shared.DescribeDomainRequest describeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n    * ListDomains returns the information and configuration for all domains.\n    **/\n    shared.ListDomainsResponse ListDomains(1: shared.ListDomainsRequest listRequest)\n      throws (\n        1: shared.BadRequestError badRequestError,\n        3: shared.EntityNotExistsError entityNotExistError,\n        4: shared.ServiceBusyError serviceBusyError,\n        5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n      )\n\n  /**\n  * UpdateDomain is used to update the information and configuration for a registered domain.\n  **/\n  shared.UpdateDomainResponse UpdateDomain(1: shared.UpdateDomainRequest updateRequest)\n      throws (\n        1: shared.BadRequestError badRequestError,\n        3: shared.EntityNotExistsError entityNotExistError,\n        4: shared.ServiceBusyError serviceBusyError,\n        5: shared.DomainNotActiveError domainNotActiveError,\n        6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n      )\n\n  /**\n  * DeprecateDomain us used to update status of a registered domain to DEPRECATED.  Once the domain is deprecated\n  * it cannot be used to start new workflow executions.  Existing workflow executions will continue to run on\n  * deprecated domains.\n  **/\n  void DeprecateDomain(1: shared.DeprecateDomainRequest deprecateRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * StartWorkflowExecution starts a new long running workflow instance.  It will create the instance with\n  * 'WorkflowExecutionStarted' event in history and also schedule the first DecisionTask for the worker to make the\n  * first decision for this instance.  It will return 'WorkflowExecutionAlreadyStartedError', if an instance already\n  * exists with same workflowId.\n  **/\n  shared.StartWorkflowExecutionResponse StartWorkflowExecution(1: shared.StartWorkflowExecutionRequest startRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.WorkflowExecutionAlreadyStartedError sessionAlreadyExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.EntityNotExistsError entityNotExistError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * Returns the history of specified workflow execution.  It fails with 'EntityNotExistError' if speficied workflow\n  * execution in unknown to the service.\n  **/\n  shared.GetWorkflowExecutionHistoryResponse GetWorkflowExecutionHistory(1: shared.GetWorkflowExecutionHistoryRequest getRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * PollForDecisionTask is called by application worker to process DecisionTask from a specific taskList.  A\n  * DecisionTask is dispatched to callers for active workflow executions, with pending decisions.\n  * Application is then expected to call 'RespondDecisionTaskCompleted' API when it is done processing the DecisionTask.\n  * It will also create a 'DecisionTaskStarted' event in the history for that session before handing off DecisionTask to\n  * application worker.\n  **/\n  shared.PollForDecisionTaskResponse PollForDecisionTask(1: shared.PollForDecisionTaskRequest pollRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.ServiceBusyError serviceBusyError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.EntityNotExistsError entityNotExistError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondDecisionTaskCompleted is called by application worker to complete a DecisionTask handed as a result of\n  * 'PollForDecisionTask' API call.  Completing a DecisionTask will result in new events for the workflow execution and\n  * potentially new ActivityTask being created for corresponding decisions.  It will also create a DecisionTaskCompleted\n  * event in the history for that session.  Use the 'taskToken' provided as response of PollForDecisionTask API call\n  * for completing the DecisionTask.\n  * The response could contain a new decision task if there is one or if the request asking for one.\n  **/\n  shared.RespondDecisionTaskCompletedResponse RespondDecisionTaskCompleted(1: shared.RespondDecisionTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondDecisionTaskFailed is called by application worker to indicate failure.  This results in\n  * DecisionTaskFailedEvent written to the history and a new DecisionTask created.  This API can be used by client to\n  * either clear sticky tasklist or report any panics during DecisionTask processing.  Cadence will only append first\n  * DecisionTaskFailed event to the history of workflow execution for consecutive failures.\n  **/\n  void RespondDecisionTaskFailed(1: shared.RespondDecisionTaskFailedRequest failedRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * PollForActivityTask is called by application worker to process ActivityTask from a specific taskList.  ActivityTask\n  * is dispatched to callers whenever a ScheduleTask decision is made for a workflow execution.\n  * Application is expected to call 'RespondActivityTaskCompleted' or 'RespondActivityTaskFailed' once it is done\n  * processing the task.\n  * Application also needs to call 'RecordActivityTaskHeartbeat' API within 'heartbeatTimeoutSeconds' interval to\n  * prevent the task from getting timed out.  An event 'ActivityTaskStarted' event is also written to workflow execution\n  * history before the ActivityTask is dispatched to application worker.\n  **/\n  shared.PollForActivityTaskResponse PollForActivityTask(1: shared.PollForActivityTaskRequest pollRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.ServiceBusyError serviceBusyError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.EntityNotExistsError entityNotExistError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RecordActivityTaskHeartbeat is called by application worker while it is processing an ActivityTask.  If worker fails\n  * to heartbeat within 'heartbeatTimeoutSeconds' interval for the ActivityTask, then it will be marked as timedout and\n  * 'ActivityTaskTimedOut' event will be written to the workflow history.  Calling 'RecordActivityTaskHeartbeat' will\n  * fail with 'EntityNotExistsError' in such situations.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for heartbeating.\n  **/\n  shared.RecordActivityTaskHeartbeatResponse RecordActivityTaskHeartbeat(1: shared.RecordActivityTaskHeartbeatRequest heartbeatRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RecordActivityTaskHeartbeatByID is called by application worker while it is processing an ActivityTask.  If worker fails\n  * to heartbeat within 'heartbeatTimeoutSeconds' interval for the ActivityTask, then it will be marked as timedout and\n  * 'ActivityTaskTimedOut' event will be written to the workflow history.  Calling 'RecordActivityTaskHeartbeatByID' will\n  * fail with 'EntityNotExistsError' in such situations.  Instead of using 'taskToken' like in RecordActivityTaskHeartbeat,\n  * use Domain, WorkflowID and ActivityID\n  **/\n  shared.RecordActivityTaskHeartbeatResponse RecordActivityTaskHeartbeatByID(1: shared.RecordActivityTaskHeartbeatByIDRequest heartbeatRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCompleted is called by application worker when it is done processing an ActivityTask.  It will\n  * result in a new 'ActivityTaskCompleted' event being written to the workflow history and a new DecisionTask\n  * created for the workflow so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskCompleted(1: shared.RespondActivityTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCompletedByID is called by application worker when it is done processing an ActivityTask.\n  * It will result in a new 'ActivityTaskCompleted' event being written to the workflow history and a new DecisionTask\n  * created for the workflow so new decisions could be made.  Similar to RespondActivityTaskCompleted but use Domain,\n  * WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskCompletedByID(1: shared.RespondActivityTaskCompletedByIDRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskFailed is called by application worker when it is done processing an ActivityTask.  It will\n  * result in a new 'ActivityTaskFailed' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskFailed(1: shared.RespondActivityTaskFailedRequest failRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskFailedByID is called by application worker when it is done processing an ActivityTask.\n  * It will result in a new 'ActivityTaskFailed' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Similar to RespondActivityTaskFailed but use\n  * Domain, WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void  RespondActivityTaskFailedByID(1: shared.RespondActivityTaskFailedByIDRequest failRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCanceled is called by application worker when it is successfully canceled an ActivityTask.  It will\n  * result in a new 'ActivityTaskCanceled' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Use the 'taskToken' provided as response of\n  * PollForActivityTask API call for completion. It fails with 'EntityNotExistsError' if the taskToken is not valid\n  * anymore due to activity timeout.\n  **/\n  void RespondActivityTaskCanceled(1: shared.RespondActivityTaskCanceledRequest canceledRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondActivityTaskCanceledByID is called by application worker when it is successfully canceled an ActivityTask.\n  * It will result in a new 'ActivityTaskCanceled' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made.  Similar to RespondActivityTaskCanceled but use\n  * Domain, WorkflowID and ActivityID instead of 'taskToken' for completion. It fails with 'EntityNotExistsError'\n  * if the these IDs are not valid anymore due to activity timeout.\n  **/\n  void RespondActivityTaskCanceledByID(1: shared.RespondActivityTaskCanceledByIDRequest canceledRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.DomainNotActiveError domainNotActiveError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RequestCancelWorkflowExecution is called by application worker when it wants to request cancellation of a workflow instance.\n  * It will result in a new 'WorkflowExecutionCancelRequested' event being written to the workflow history and a new DecisionTask\n  * created for the workflow instance so new decisions could be made. It fails with 'EntityNotExistsError' if the workflow is not valid\n  * anymore due to completion or doesn't exist.\n  **/\n  void RequestCancelWorkflowExecution(1: shared.RequestCancelWorkflowExecutionRequest cancelRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.CancellationAlreadyRequestedError cancellationAlreadyRequestedError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.LimitExceededError limitExceededError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * SignalWorkflowExecution is used to send a signal event to running workflow execution.  This results in\n  * WorkflowExecutionSignaled event recorded in the history and a decision task being created for the execution.\n  **/\n  void SignalWorkflowExecution(1: shared.SignalWorkflowExecutionRequest signalRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * SignalWithStartWorkflowExecution is used to ensure sending signal to a workflow.\n  * If the workflow is running, this results in WorkflowExecutionSignaled event being recorded in the history\n  * and a decision task being created for the execution.\n  * If the workflow is not running or not found, this results in WorkflowExecutionStarted and WorkflowExecutionSignaled\n  * events being recorded in history, and a decision task being created for the execution\n  **/\n  shared.StartWorkflowExecutionResponse SignalWithStartWorkflowExecution(1: shared.SignalWithStartWorkflowExecutionRequest signalWithStartRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.WorkflowExecutionAlreadyStartedError workflowAlreadyStartedError,\n      8: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n    * ResetWorkflowExecution reset an existing workflow execution to DecisionTaskCompleted event(exclusive).\n    * And it will immediately terminating the current execution instance.\n    **/\n  shared.ResetWorkflowExecutionResponse ResetWorkflowExecution(1: shared.ResetWorkflowExecutionRequest resetRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * TerminateWorkflowExecution terminates an existing workflow execution by recording WorkflowExecutionTerminated event\n  * in the history and immediately terminating the execution instance.\n  **/\n  void TerminateWorkflowExecution(1: shared.TerminateWorkflowExecutionRequest terminateRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.DomainNotActiveError domainNotActiveError,\n      6: shared.LimitExceededError limitExceededError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListOpenWorkflowExecutions is a visibility API to list the open executions in a specific domain.\n  **/\n  shared.ListOpenWorkflowExecutionsResponse ListOpenWorkflowExecutions(1: shared.ListOpenWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.LimitExceededError limitExceededError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListClosedWorkflowExecutions is a visibility API to list the closed executions in a specific domain.\n  **/\n  shared.ListClosedWorkflowExecutionsResponse ListClosedWorkflowExecutions(1: shared.ListClosedWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListWorkflowExecutions is a visibility API to list workflow executions in a specific domain.\n  **/\n  shared.ListWorkflowExecutionsResponse ListWorkflowExecutions(1: shared.ListWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ListArchivedWorkflowExecutions is a visibility API to list archived workflow executions in a specific domain.\n  **/\n  shared.ListArchivedWorkflowExecutionsResponse ListArchivedWorkflowExecutions(1: shared.ListArchivedWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * ScanWorkflowExecutions is a visibility API to list large amount of workflow executions in a specific domain without order.\n  **/\n  shared.ListWorkflowExecutionsResponse ScanWorkflowExecutions(1: shared.ListWorkflowExecutionsRequest listRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * CountWorkflowExecutions is a visibility API to count of workflow executions in a specific domain.\n  **/\n  shared.CountWorkflowExecutionsResponse CountWorkflowExecutions(1: shared.CountWorkflowExecutionsRequest countRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.ServiceBusyError serviceBusyError,\n      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * GetSearchAttributes is a visibility API to get all legal keys that could be used in list APIs\n  **/\n  shared.GetSearchAttributesResponse GetSearchAttributes()\n    throws (\n      2: shared.ServiceBusyError serviceBusyError,\n      3: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * RespondQueryTaskCompleted is called by application worker to complete a QueryTask (which is a DecisionTask for query)\n  * as a result of 'PollForDecisionTask' API call. Completing a QueryTask will unblock the client call to 'QueryWorkflow'\n  * API and return the query result to client as a response to 'QueryWorkflow' API call.\n  **/\n  void RespondQueryTaskCompleted(1: shared.RespondQueryTaskCompletedRequest completeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * Reset the sticky tasklist related information in mutable state of a given workflow.\n  * Things cleared are:\n  * 1. StickyTaskList\n  * 2. StickyScheduleToStartTimeout\n  * 3. ClientLibraryVersion\n  * 4. ClientFeatureVersion\n  * 5. ClientImpl\n  **/\n  shared.ResetStickyTaskListResponse ResetStickyTaskList(1: shared.ResetStickyTaskListRequest resetRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.DomainNotActiveError domainNotActiveError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * QueryWorkflow returns query result for a specified workflow execution\n  **/\n  shared.QueryWorkflowResponse QueryWorkflow(1: shared.QueryWorkflowRequest queryRequest)\n\tthrows (\n\t  1: shared.BadRequestError badRequestError,\n\t  3: shared.EntityNotExistsError entityNotExistError,\n\t  4: shared.QueryFailedError queryFailedError,\n\t  5: shared.LimitExceededError limitExceededError,\n      6: shared.ServiceBusyError serviceBusyError,\n      7: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n\t)\n\n  /**\n  * DescribeWorkflowExecution returns information about the specified workflow execution.\n  **/\n  shared.DescribeWorkflowExecutionResponse DescribeWorkflowExecution(1: shared.DescribeWorkflowExecutionRequest describeRequest)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * DescribeTaskList returns information about the target tasklist, right now this API returns the\n  * pollers which polled this tasklist in last few minutes.\n  **/\n  shared.DescribeTaskListResponse DescribeTaskList(1: shared.DescribeTaskListRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n      6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,\n    )\n\n  /**\n  * GetClusterInfo returns information about cadence cluster\n  **/\n  shared.ClusterInfo GetClusterInfo()\n    throws (\n      1: shared.InternalServiceError internalServiceError,\n      2: shared.ServiceBusyError serviceBusyError,\n    )\n\n   /**\n   * ReapplyEvents applies stale events to the current workflow and current run\n   **/\n  shared.ListTaskListPartitionsResponse ListTaskListPartitions(1: shared.ListTaskListPartitionsRequest request)\n    throws (\n      1: shared.BadRequestError badRequestError,\n      3: shared.EntityNotExistsError entityNotExistError,\n      4: shared.LimitExceededError limitExceededError,\n      5: shared.ServiceBusyError serviceBusyError,\n    )\n}\n"
 
 // WorkflowService_CountWorkflowExecutions_Args represents the arguments for the WorkflowService.CountWorkflowExecutions function.
 //
@@ -3887,1292 +3887,6 @@ func (v *WorkflowService_GetClusterInfo_Result) MethodName() string {
 //
 // This will always be Reply for this struct.
 func (v *WorkflowService_GetClusterInfo_Result) EnvelopeType() wire.EnvelopeType {
-	return wire.Reply
-}
-
-// WorkflowService_GetDomainReplicationMessages_Args represents the arguments for the WorkflowService.GetDomainReplicationMessages function.
-//
-// The arguments for GetDomainReplicationMessages are sent and received over the wire as this struct.
-type WorkflowService_GetDomainReplicationMessages_Args struct {
-	Request *replicator.GetDomainReplicationMessagesRequest `json:"request,omitempty"`
-}
-
-// ToWire translates a WorkflowService_GetDomainReplicationMessages_Args struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_GetDomainReplicationMessages_Args) ToWire() (wire.Value, error) {
-	var (
-		fields [1]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.Request != nil {
-		w, err = v.Request.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-func _GetDomainReplicationMessagesRequest_Read(w wire.Value) (*replicator.GetDomainReplicationMessagesRequest, error) {
-	var v replicator.GetDomainReplicationMessagesRequest
-	err := v.FromWire(w)
-	return &v, err
-}
-
-// FromWire deserializes a WorkflowService_GetDomainReplicationMessages_Args struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_GetDomainReplicationMessages_Args struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_GetDomainReplicationMessages_Args
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_GetDomainReplicationMessages_Args) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.Request, err = _GetDomainReplicationMessagesRequest_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_GetDomainReplicationMessages_Args
-// struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [1]string
-	i := 0
-	if v.Request != nil {
-		fields[i] = fmt.Sprintf("Request: %v", v.Request)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_GetDomainReplicationMessages_Args{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_GetDomainReplicationMessages_Args match the
-// provided WorkflowService_GetDomainReplicationMessages_Args.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) Equals(rhs *WorkflowService_GetDomainReplicationMessages_Args) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.Request == nil && rhs.Request == nil) || (v.Request != nil && rhs.Request != nil && v.Request.Equals(rhs.Request))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_GetDomainReplicationMessages_Args.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.Request != nil {
-		err = multierr.Append(err, enc.AddObject("request", v.Request))
-	}
-	return err
-}
-
-// GetRequest returns the value of Request if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) GetRequest() (o *replicator.GetDomainReplicationMessagesRequest) {
-	if v != nil && v.Request != nil {
-		return v.Request
-	}
-
-	return
-}
-
-// IsSetRequest returns true if Request is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) IsSetRequest() bool {
-	return v != nil && v.Request != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the arguments.
-//
-// This will always be "GetDomainReplicationMessages" for this struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) MethodName() string {
-	return "GetDomainReplicationMessages"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Call for this struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Args) EnvelopeType() wire.EnvelopeType {
-	return wire.Call
-}
-
-// WorkflowService_GetDomainReplicationMessages_Helper provides functions that aid in handling the
-// parameters and return values of the WorkflowService.GetDomainReplicationMessages
-// function.
-var WorkflowService_GetDomainReplicationMessages_Helper = struct {
-	// Args accepts the parameters of GetDomainReplicationMessages in-order and returns
-	// the arguments struct for the function.
-	Args func(
-		request *replicator.GetDomainReplicationMessagesRequest,
-	) *WorkflowService_GetDomainReplicationMessages_Args
-
-	// IsException returns true if the given error can be thrown
-	// by GetDomainReplicationMessages.
-	//
-	// An error can be thrown by GetDomainReplicationMessages only if the
-	// corresponding exception type was mentioned in the 'throws'
-	// section for it in the Thrift file.
-	IsException func(error) bool
-
-	// WrapResponse returns the result struct for GetDomainReplicationMessages
-	// given its return value and error.
-	//
-	// This allows mapping values and errors returned by
-	// GetDomainReplicationMessages into a serializable result struct.
-	// WrapResponse returns a non-nil error if the provided
-	// error cannot be thrown by GetDomainReplicationMessages
-	//
-	//   value, err := GetDomainReplicationMessages(args)
-	//   result, err := WorkflowService_GetDomainReplicationMessages_Helper.WrapResponse(value, err)
-	//   if err != nil {
-	//     return fmt.Errorf("unexpected error from GetDomainReplicationMessages: %v", err)
-	//   }
-	//   serialize(result)
-	WrapResponse func(*replicator.GetDomainReplicationMessagesResponse, error) (*WorkflowService_GetDomainReplicationMessages_Result, error)
-
-	// UnwrapResponse takes the result struct for GetDomainReplicationMessages
-	// and returns the value or error returned by it.
-	//
-	// The error is non-nil only if GetDomainReplicationMessages threw an
-	// exception.
-	//
-	//   result := deserialize(bytes)
-	//   value, err := WorkflowService_GetDomainReplicationMessages_Helper.UnwrapResponse(result)
-	UnwrapResponse func(*WorkflowService_GetDomainReplicationMessages_Result) (*replicator.GetDomainReplicationMessagesResponse, error)
-}{}
-
-func init() {
-	WorkflowService_GetDomainReplicationMessages_Helper.Args = func(
-		request *replicator.GetDomainReplicationMessagesRequest,
-	) *WorkflowService_GetDomainReplicationMessages_Args {
-		return &WorkflowService_GetDomainReplicationMessages_Args{
-			Request: request,
-		}
-	}
-
-	WorkflowService_GetDomainReplicationMessages_Helper.IsException = func(err error) bool {
-		switch err.(type) {
-		case *shared.BadRequestError:
-			return true
-		case *shared.LimitExceededError:
-			return true
-		case *shared.ServiceBusyError:
-			return true
-		case *shared.ClientVersionNotSupportedError:
-			return true
-		default:
-			return false
-		}
-	}
-
-	WorkflowService_GetDomainReplicationMessages_Helper.WrapResponse = func(success *replicator.GetDomainReplicationMessagesResponse, err error) (*WorkflowService_GetDomainReplicationMessages_Result, error) {
-		if err == nil {
-			return &WorkflowService_GetDomainReplicationMessages_Result{Success: success}, nil
-		}
-
-		switch e := err.(type) {
-		case *shared.BadRequestError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetDomainReplicationMessages_Result.BadRequestError")
-			}
-			return &WorkflowService_GetDomainReplicationMessages_Result{BadRequestError: e}, nil
-		case *shared.LimitExceededError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetDomainReplicationMessages_Result.LimitExceededError")
-			}
-			return &WorkflowService_GetDomainReplicationMessages_Result{LimitExceededError: e}, nil
-		case *shared.ServiceBusyError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetDomainReplicationMessages_Result.ServiceBusyError")
-			}
-			return &WorkflowService_GetDomainReplicationMessages_Result{ServiceBusyError: e}, nil
-		case *shared.ClientVersionNotSupportedError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetDomainReplicationMessages_Result.ClientVersionNotSupportedError")
-			}
-			return &WorkflowService_GetDomainReplicationMessages_Result{ClientVersionNotSupportedError: e}, nil
-		}
-
-		return nil, err
-	}
-	WorkflowService_GetDomainReplicationMessages_Helper.UnwrapResponse = func(result *WorkflowService_GetDomainReplicationMessages_Result) (success *replicator.GetDomainReplicationMessagesResponse, err error) {
-		if result.BadRequestError != nil {
-			err = result.BadRequestError
-			return
-		}
-		if result.LimitExceededError != nil {
-			err = result.LimitExceededError
-			return
-		}
-		if result.ServiceBusyError != nil {
-			err = result.ServiceBusyError
-			return
-		}
-		if result.ClientVersionNotSupportedError != nil {
-			err = result.ClientVersionNotSupportedError
-			return
-		}
-
-		if result.Success != nil {
-			success = result.Success
-			return
-		}
-
-		err = errors.New("expected a non-void result")
-		return
-	}
-
-}
-
-// WorkflowService_GetDomainReplicationMessages_Result represents the result of a WorkflowService.GetDomainReplicationMessages function call.
-//
-// The result of a GetDomainReplicationMessages execution is sent and received over the wire as this struct.
-//
-// Success is set only if the function did not throw an exception.
-type WorkflowService_GetDomainReplicationMessages_Result struct {
-	// Value returned by GetDomainReplicationMessages after a successful execution.
-	Success                        *replicator.GetDomainReplicationMessagesResponse `json:"success,omitempty"`
-	BadRequestError                *shared.BadRequestError                          `json:"badRequestError,omitempty"`
-	LimitExceededError             *shared.LimitExceededError                       `json:"limitExceededError,omitempty"`
-	ServiceBusyError               *shared.ServiceBusyError                         `json:"serviceBusyError,omitempty"`
-	ClientVersionNotSupportedError *shared.ClientVersionNotSupportedError           `json:"clientVersionNotSupportedError,omitempty"`
-}
-
-// ToWire translates a WorkflowService_GetDomainReplicationMessages_Result struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_GetDomainReplicationMessages_Result) ToWire() (wire.Value, error) {
-	var (
-		fields [5]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.Success != nil {
-		w, err = v.Success.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 0, Value: w}
-		i++
-	}
-	if v.BadRequestError != nil {
-		w, err = v.BadRequestError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-	if v.LimitExceededError != nil {
-		w, err = v.LimitExceededError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 3, Value: w}
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		w, err = v.ServiceBusyError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 4, Value: w}
-		i++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		w, err = v.ClientVersionNotSupportedError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 5, Value: w}
-		i++
-	}
-
-	if i != 1 {
-		return wire.Value{}, fmt.Errorf("WorkflowService_GetDomainReplicationMessages_Result should have exactly one field: got %v fields", i)
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-func _GetDomainReplicationMessagesResponse_Read(w wire.Value) (*replicator.GetDomainReplicationMessagesResponse, error) {
-	var v replicator.GetDomainReplicationMessagesResponse
-	err := v.FromWire(w)
-	return &v, err
-}
-
-// FromWire deserializes a WorkflowService_GetDomainReplicationMessages_Result struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_GetDomainReplicationMessages_Result struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_GetDomainReplicationMessages_Result
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_GetDomainReplicationMessages_Result) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 0:
-			if field.Value.Type() == wire.TStruct {
-				v.Success, err = _GetDomainReplicationMessagesResponse_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.BadRequestError, err = _BadRequestError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 3:
-			if field.Value.Type() == wire.TStruct {
-				v.LimitExceededError, err = _LimitExceededError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 4:
-			if field.Value.Type() == wire.TStruct {
-				v.ServiceBusyError, err = _ServiceBusyError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 5:
-			if field.Value.Type() == wire.TStruct {
-				v.ClientVersionNotSupportedError, err = _ClientVersionNotSupportedError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	count := 0
-	if v.Success != nil {
-		count++
-	}
-	if v.BadRequestError != nil {
-		count++
-	}
-	if v.LimitExceededError != nil {
-		count++
-	}
-	if v.ServiceBusyError != nil {
-		count++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		count++
-	}
-	if count != 1 {
-		return fmt.Errorf("WorkflowService_GetDomainReplicationMessages_Result should have exactly one field: got %v fields", count)
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_GetDomainReplicationMessages_Result
-// struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [5]string
-	i := 0
-	if v.Success != nil {
-		fields[i] = fmt.Sprintf("Success: %v", v.Success)
-		i++
-	}
-	if v.BadRequestError != nil {
-		fields[i] = fmt.Sprintf("BadRequestError: %v", v.BadRequestError)
-		i++
-	}
-	if v.LimitExceededError != nil {
-		fields[i] = fmt.Sprintf("LimitExceededError: %v", v.LimitExceededError)
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		fields[i] = fmt.Sprintf("ServiceBusyError: %v", v.ServiceBusyError)
-		i++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		fields[i] = fmt.Sprintf("ClientVersionNotSupportedError: %v", v.ClientVersionNotSupportedError)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_GetDomainReplicationMessages_Result{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_GetDomainReplicationMessages_Result match the
-// provided WorkflowService_GetDomainReplicationMessages_Result.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) Equals(rhs *WorkflowService_GetDomainReplicationMessages_Result) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.Success == nil && rhs.Success == nil) || (v.Success != nil && rhs.Success != nil && v.Success.Equals(rhs.Success))) {
-		return false
-	}
-	if !((v.BadRequestError == nil && rhs.BadRequestError == nil) || (v.BadRequestError != nil && rhs.BadRequestError != nil && v.BadRequestError.Equals(rhs.BadRequestError))) {
-		return false
-	}
-	if !((v.LimitExceededError == nil && rhs.LimitExceededError == nil) || (v.LimitExceededError != nil && rhs.LimitExceededError != nil && v.LimitExceededError.Equals(rhs.LimitExceededError))) {
-		return false
-	}
-	if !((v.ServiceBusyError == nil && rhs.ServiceBusyError == nil) || (v.ServiceBusyError != nil && rhs.ServiceBusyError != nil && v.ServiceBusyError.Equals(rhs.ServiceBusyError))) {
-		return false
-	}
-	if !((v.ClientVersionNotSupportedError == nil && rhs.ClientVersionNotSupportedError == nil) || (v.ClientVersionNotSupportedError != nil && rhs.ClientVersionNotSupportedError != nil && v.ClientVersionNotSupportedError.Equals(rhs.ClientVersionNotSupportedError))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_GetDomainReplicationMessages_Result.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.Success != nil {
-		err = multierr.Append(err, enc.AddObject("success", v.Success))
-	}
-	if v.BadRequestError != nil {
-		err = multierr.Append(err, enc.AddObject("badRequestError", v.BadRequestError))
-	}
-	if v.LimitExceededError != nil {
-		err = multierr.Append(err, enc.AddObject("limitExceededError", v.LimitExceededError))
-	}
-	if v.ServiceBusyError != nil {
-		err = multierr.Append(err, enc.AddObject("serviceBusyError", v.ServiceBusyError))
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		err = multierr.Append(err, enc.AddObject("clientVersionNotSupportedError", v.ClientVersionNotSupportedError))
-	}
-	return err
-}
-
-// GetSuccess returns the value of Success if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) GetSuccess() (o *replicator.GetDomainReplicationMessagesResponse) {
-	if v != nil && v.Success != nil {
-		return v.Success
-	}
-
-	return
-}
-
-// IsSetSuccess returns true if Success is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) IsSetSuccess() bool {
-	return v != nil && v.Success != nil
-}
-
-// GetBadRequestError returns the value of BadRequestError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) GetBadRequestError() (o *shared.BadRequestError) {
-	if v != nil && v.BadRequestError != nil {
-		return v.BadRequestError
-	}
-
-	return
-}
-
-// IsSetBadRequestError returns true if BadRequestError is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) IsSetBadRequestError() bool {
-	return v != nil && v.BadRequestError != nil
-}
-
-// GetLimitExceededError returns the value of LimitExceededError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) GetLimitExceededError() (o *shared.LimitExceededError) {
-	if v != nil && v.LimitExceededError != nil {
-		return v.LimitExceededError
-	}
-
-	return
-}
-
-// IsSetLimitExceededError returns true if LimitExceededError is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) IsSetLimitExceededError() bool {
-	return v != nil && v.LimitExceededError != nil
-}
-
-// GetServiceBusyError returns the value of ServiceBusyError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) GetServiceBusyError() (o *shared.ServiceBusyError) {
-	if v != nil && v.ServiceBusyError != nil {
-		return v.ServiceBusyError
-	}
-
-	return
-}
-
-// IsSetServiceBusyError returns true if ServiceBusyError is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) IsSetServiceBusyError() bool {
-	return v != nil && v.ServiceBusyError != nil
-}
-
-// GetClientVersionNotSupportedError returns the value of ClientVersionNotSupportedError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) GetClientVersionNotSupportedError() (o *shared.ClientVersionNotSupportedError) {
-	if v != nil && v.ClientVersionNotSupportedError != nil {
-		return v.ClientVersionNotSupportedError
-	}
-
-	return
-}
-
-// IsSetClientVersionNotSupportedError returns true if ClientVersionNotSupportedError is not nil.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) IsSetClientVersionNotSupportedError() bool {
-	return v != nil && v.ClientVersionNotSupportedError != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the result.
-//
-// This will always be "GetDomainReplicationMessages" for this struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) MethodName() string {
-	return "GetDomainReplicationMessages"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Reply for this struct.
-func (v *WorkflowService_GetDomainReplicationMessages_Result) EnvelopeType() wire.EnvelopeType {
-	return wire.Reply
-}
-
-// WorkflowService_GetReplicationMessages_Args represents the arguments for the WorkflowService.GetReplicationMessages function.
-//
-// The arguments for GetReplicationMessages are sent and received over the wire as this struct.
-type WorkflowService_GetReplicationMessages_Args struct {
-	Request *replicator.GetReplicationMessagesRequest `json:"request,omitempty"`
-}
-
-// ToWire translates a WorkflowService_GetReplicationMessages_Args struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_GetReplicationMessages_Args) ToWire() (wire.Value, error) {
-	var (
-		fields [1]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.Request != nil {
-		w, err = v.Request.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-func _GetReplicationMessagesRequest_Read(w wire.Value) (*replicator.GetReplicationMessagesRequest, error) {
-	var v replicator.GetReplicationMessagesRequest
-	err := v.FromWire(w)
-	return &v, err
-}
-
-// FromWire deserializes a WorkflowService_GetReplicationMessages_Args struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_GetReplicationMessages_Args struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_GetReplicationMessages_Args
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_GetReplicationMessages_Args) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.Request, err = _GetReplicationMessagesRequest_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_GetReplicationMessages_Args
-// struct.
-func (v *WorkflowService_GetReplicationMessages_Args) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [1]string
-	i := 0
-	if v.Request != nil {
-		fields[i] = fmt.Sprintf("Request: %v", v.Request)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_GetReplicationMessages_Args{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_GetReplicationMessages_Args match the
-// provided WorkflowService_GetReplicationMessages_Args.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_GetReplicationMessages_Args) Equals(rhs *WorkflowService_GetReplicationMessages_Args) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.Request == nil && rhs.Request == nil) || (v.Request != nil && rhs.Request != nil && v.Request.Equals(rhs.Request))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_GetReplicationMessages_Args.
-func (v *WorkflowService_GetReplicationMessages_Args) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.Request != nil {
-		err = multierr.Append(err, enc.AddObject("request", v.Request))
-	}
-	return err
-}
-
-// GetRequest returns the value of Request if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Args) GetRequest() (o *replicator.GetReplicationMessagesRequest) {
-	if v != nil && v.Request != nil {
-		return v.Request
-	}
-
-	return
-}
-
-// IsSetRequest returns true if Request is not nil.
-func (v *WorkflowService_GetReplicationMessages_Args) IsSetRequest() bool {
-	return v != nil && v.Request != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the arguments.
-//
-// This will always be "GetReplicationMessages" for this struct.
-func (v *WorkflowService_GetReplicationMessages_Args) MethodName() string {
-	return "GetReplicationMessages"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Call for this struct.
-func (v *WorkflowService_GetReplicationMessages_Args) EnvelopeType() wire.EnvelopeType {
-	return wire.Call
-}
-
-// WorkflowService_GetReplicationMessages_Helper provides functions that aid in handling the
-// parameters and return values of the WorkflowService.GetReplicationMessages
-// function.
-var WorkflowService_GetReplicationMessages_Helper = struct {
-	// Args accepts the parameters of GetReplicationMessages in-order and returns
-	// the arguments struct for the function.
-	Args func(
-		request *replicator.GetReplicationMessagesRequest,
-	) *WorkflowService_GetReplicationMessages_Args
-
-	// IsException returns true if the given error can be thrown
-	// by GetReplicationMessages.
-	//
-	// An error can be thrown by GetReplicationMessages only if the
-	// corresponding exception type was mentioned in the 'throws'
-	// section for it in the Thrift file.
-	IsException func(error) bool
-
-	// WrapResponse returns the result struct for GetReplicationMessages
-	// given its return value and error.
-	//
-	// This allows mapping values and errors returned by
-	// GetReplicationMessages into a serializable result struct.
-	// WrapResponse returns a non-nil error if the provided
-	// error cannot be thrown by GetReplicationMessages
-	//
-	//   value, err := GetReplicationMessages(args)
-	//   result, err := WorkflowService_GetReplicationMessages_Helper.WrapResponse(value, err)
-	//   if err != nil {
-	//     return fmt.Errorf("unexpected error from GetReplicationMessages: %v", err)
-	//   }
-	//   serialize(result)
-	WrapResponse func(*replicator.GetReplicationMessagesResponse, error) (*WorkflowService_GetReplicationMessages_Result, error)
-
-	// UnwrapResponse takes the result struct for GetReplicationMessages
-	// and returns the value or error returned by it.
-	//
-	// The error is non-nil only if GetReplicationMessages threw an
-	// exception.
-	//
-	//   result := deserialize(bytes)
-	//   value, err := WorkflowService_GetReplicationMessages_Helper.UnwrapResponse(result)
-	UnwrapResponse func(*WorkflowService_GetReplicationMessages_Result) (*replicator.GetReplicationMessagesResponse, error)
-}{}
-
-func init() {
-	WorkflowService_GetReplicationMessages_Helper.Args = func(
-		request *replicator.GetReplicationMessagesRequest,
-	) *WorkflowService_GetReplicationMessages_Args {
-		return &WorkflowService_GetReplicationMessages_Args{
-			Request: request,
-		}
-	}
-
-	WorkflowService_GetReplicationMessages_Helper.IsException = func(err error) bool {
-		switch err.(type) {
-		case *shared.BadRequestError:
-			return true
-		case *shared.LimitExceededError:
-			return true
-		case *shared.ServiceBusyError:
-			return true
-		case *shared.ClientVersionNotSupportedError:
-			return true
-		default:
-			return false
-		}
-	}
-
-	WorkflowService_GetReplicationMessages_Helper.WrapResponse = func(success *replicator.GetReplicationMessagesResponse, err error) (*WorkflowService_GetReplicationMessages_Result, error) {
-		if err == nil {
-			return &WorkflowService_GetReplicationMessages_Result{Success: success}, nil
-		}
-
-		switch e := err.(type) {
-		case *shared.BadRequestError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetReplicationMessages_Result.BadRequestError")
-			}
-			return &WorkflowService_GetReplicationMessages_Result{BadRequestError: e}, nil
-		case *shared.LimitExceededError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetReplicationMessages_Result.LimitExceededError")
-			}
-			return &WorkflowService_GetReplicationMessages_Result{LimitExceededError: e}, nil
-		case *shared.ServiceBusyError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetReplicationMessages_Result.ServiceBusyError")
-			}
-			return &WorkflowService_GetReplicationMessages_Result{ServiceBusyError: e}, nil
-		case *shared.ClientVersionNotSupportedError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_GetReplicationMessages_Result.ClientVersionNotSupportedError")
-			}
-			return &WorkflowService_GetReplicationMessages_Result{ClientVersionNotSupportedError: e}, nil
-		}
-
-		return nil, err
-	}
-	WorkflowService_GetReplicationMessages_Helper.UnwrapResponse = func(result *WorkflowService_GetReplicationMessages_Result) (success *replicator.GetReplicationMessagesResponse, err error) {
-		if result.BadRequestError != nil {
-			err = result.BadRequestError
-			return
-		}
-		if result.LimitExceededError != nil {
-			err = result.LimitExceededError
-			return
-		}
-		if result.ServiceBusyError != nil {
-			err = result.ServiceBusyError
-			return
-		}
-		if result.ClientVersionNotSupportedError != nil {
-			err = result.ClientVersionNotSupportedError
-			return
-		}
-
-		if result.Success != nil {
-			success = result.Success
-			return
-		}
-
-		err = errors.New("expected a non-void result")
-		return
-	}
-
-}
-
-// WorkflowService_GetReplicationMessages_Result represents the result of a WorkflowService.GetReplicationMessages function call.
-//
-// The result of a GetReplicationMessages execution is sent and received over the wire as this struct.
-//
-// Success is set only if the function did not throw an exception.
-type WorkflowService_GetReplicationMessages_Result struct {
-	// Value returned by GetReplicationMessages after a successful execution.
-	Success                        *replicator.GetReplicationMessagesResponse `json:"success,omitempty"`
-	BadRequestError                *shared.BadRequestError                    `json:"badRequestError,omitempty"`
-	LimitExceededError             *shared.LimitExceededError                 `json:"limitExceededError,omitempty"`
-	ServiceBusyError               *shared.ServiceBusyError                   `json:"serviceBusyError,omitempty"`
-	ClientVersionNotSupportedError *shared.ClientVersionNotSupportedError     `json:"clientVersionNotSupportedError,omitempty"`
-}
-
-// ToWire translates a WorkflowService_GetReplicationMessages_Result struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_GetReplicationMessages_Result) ToWire() (wire.Value, error) {
-	var (
-		fields [5]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.Success != nil {
-		w, err = v.Success.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 0, Value: w}
-		i++
-	}
-	if v.BadRequestError != nil {
-		w, err = v.BadRequestError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-	if v.LimitExceededError != nil {
-		w, err = v.LimitExceededError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 3, Value: w}
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		w, err = v.ServiceBusyError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 4, Value: w}
-		i++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		w, err = v.ClientVersionNotSupportedError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 5, Value: w}
-		i++
-	}
-
-	if i != 1 {
-		return wire.Value{}, fmt.Errorf("WorkflowService_GetReplicationMessages_Result should have exactly one field: got %v fields", i)
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-func _GetReplicationMessagesResponse_Read(w wire.Value) (*replicator.GetReplicationMessagesResponse, error) {
-	var v replicator.GetReplicationMessagesResponse
-	err := v.FromWire(w)
-	return &v, err
-}
-
-// FromWire deserializes a WorkflowService_GetReplicationMessages_Result struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_GetReplicationMessages_Result struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_GetReplicationMessages_Result
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_GetReplicationMessages_Result) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 0:
-			if field.Value.Type() == wire.TStruct {
-				v.Success, err = _GetReplicationMessagesResponse_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.BadRequestError, err = _BadRequestError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 3:
-			if field.Value.Type() == wire.TStruct {
-				v.LimitExceededError, err = _LimitExceededError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 4:
-			if field.Value.Type() == wire.TStruct {
-				v.ServiceBusyError, err = _ServiceBusyError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 5:
-			if field.Value.Type() == wire.TStruct {
-				v.ClientVersionNotSupportedError, err = _ClientVersionNotSupportedError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	count := 0
-	if v.Success != nil {
-		count++
-	}
-	if v.BadRequestError != nil {
-		count++
-	}
-	if v.LimitExceededError != nil {
-		count++
-	}
-	if v.ServiceBusyError != nil {
-		count++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		count++
-	}
-	if count != 1 {
-		return fmt.Errorf("WorkflowService_GetReplicationMessages_Result should have exactly one field: got %v fields", count)
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_GetReplicationMessages_Result
-// struct.
-func (v *WorkflowService_GetReplicationMessages_Result) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [5]string
-	i := 0
-	if v.Success != nil {
-		fields[i] = fmt.Sprintf("Success: %v", v.Success)
-		i++
-	}
-	if v.BadRequestError != nil {
-		fields[i] = fmt.Sprintf("BadRequestError: %v", v.BadRequestError)
-		i++
-	}
-	if v.LimitExceededError != nil {
-		fields[i] = fmt.Sprintf("LimitExceededError: %v", v.LimitExceededError)
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		fields[i] = fmt.Sprintf("ServiceBusyError: %v", v.ServiceBusyError)
-		i++
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		fields[i] = fmt.Sprintf("ClientVersionNotSupportedError: %v", v.ClientVersionNotSupportedError)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_GetReplicationMessages_Result{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_GetReplicationMessages_Result match the
-// provided WorkflowService_GetReplicationMessages_Result.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_GetReplicationMessages_Result) Equals(rhs *WorkflowService_GetReplicationMessages_Result) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.Success == nil && rhs.Success == nil) || (v.Success != nil && rhs.Success != nil && v.Success.Equals(rhs.Success))) {
-		return false
-	}
-	if !((v.BadRequestError == nil && rhs.BadRequestError == nil) || (v.BadRequestError != nil && rhs.BadRequestError != nil && v.BadRequestError.Equals(rhs.BadRequestError))) {
-		return false
-	}
-	if !((v.LimitExceededError == nil && rhs.LimitExceededError == nil) || (v.LimitExceededError != nil && rhs.LimitExceededError != nil && v.LimitExceededError.Equals(rhs.LimitExceededError))) {
-		return false
-	}
-	if !((v.ServiceBusyError == nil && rhs.ServiceBusyError == nil) || (v.ServiceBusyError != nil && rhs.ServiceBusyError != nil && v.ServiceBusyError.Equals(rhs.ServiceBusyError))) {
-		return false
-	}
-	if !((v.ClientVersionNotSupportedError == nil && rhs.ClientVersionNotSupportedError == nil) || (v.ClientVersionNotSupportedError != nil && rhs.ClientVersionNotSupportedError != nil && v.ClientVersionNotSupportedError.Equals(rhs.ClientVersionNotSupportedError))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_GetReplicationMessages_Result.
-func (v *WorkflowService_GetReplicationMessages_Result) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.Success != nil {
-		err = multierr.Append(err, enc.AddObject("success", v.Success))
-	}
-	if v.BadRequestError != nil {
-		err = multierr.Append(err, enc.AddObject("badRequestError", v.BadRequestError))
-	}
-	if v.LimitExceededError != nil {
-		err = multierr.Append(err, enc.AddObject("limitExceededError", v.LimitExceededError))
-	}
-	if v.ServiceBusyError != nil {
-		err = multierr.Append(err, enc.AddObject("serviceBusyError", v.ServiceBusyError))
-	}
-	if v.ClientVersionNotSupportedError != nil {
-		err = multierr.Append(err, enc.AddObject("clientVersionNotSupportedError", v.ClientVersionNotSupportedError))
-	}
-	return err
-}
-
-// GetSuccess returns the value of Success if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Result) GetSuccess() (o *replicator.GetReplicationMessagesResponse) {
-	if v != nil && v.Success != nil {
-		return v.Success
-	}
-
-	return
-}
-
-// IsSetSuccess returns true if Success is not nil.
-func (v *WorkflowService_GetReplicationMessages_Result) IsSetSuccess() bool {
-	return v != nil && v.Success != nil
-}
-
-// GetBadRequestError returns the value of BadRequestError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Result) GetBadRequestError() (o *shared.BadRequestError) {
-	if v != nil && v.BadRequestError != nil {
-		return v.BadRequestError
-	}
-
-	return
-}
-
-// IsSetBadRequestError returns true if BadRequestError is not nil.
-func (v *WorkflowService_GetReplicationMessages_Result) IsSetBadRequestError() bool {
-	return v != nil && v.BadRequestError != nil
-}
-
-// GetLimitExceededError returns the value of LimitExceededError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Result) GetLimitExceededError() (o *shared.LimitExceededError) {
-	if v != nil && v.LimitExceededError != nil {
-		return v.LimitExceededError
-	}
-
-	return
-}
-
-// IsSetLimitExceededError returns true if LimitExceededError is not nil.
-func (v *WorkflowService_GetReplicationMessages_Result) IsSetLimitExceededError() bool {
-	return v != nil && v.LimitExceededError != nil
-}
-
-// GetServiceBusyError returns the value of ServiceBusyError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Result) GetServiceBusyError() (o *shared.ServiceBusyError) {
-	if v != nil && v.ServiceBusyError != nil {
-		return v.ServiceBusyError
-	}
-
-	return
-}
-
-// IsSetServiceBusyError returns true if ServiceBusyError is not nil.
-func (v *WorkflowService_GetReplicationMessages_Result) IsSetServiceBusyError() bool {
-	return v != nil && v.ServiceBusyError != nil
-}
-
-// GetClientVersionNotSupportedError returns the value of ClientVersionNotSupportedError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_GetReplicationMessages_Result) GetClientVersionNotSupportedError() (o *shared.ClientVersionNotSupportedError) {
-	if v != nil && v.ClientVersionNotSupportedError != nil {
-		return v.ClientVersionNotSupportedError
-	}
-
-	return
-}
-
-// IsSetClientVersionNotSupportedError returns true if ClientVersionNotSupportedError is not nil.
-func (v *WorkflowService_GetReplicationMessages_Result) IsSetClientVersionNotSupportedError() bool {
-	return v != nil && v.ClientVersionNotSupportedError != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the result.
-//
-// This will always be "GetReplicationMessages" for this struct.
-func (v *WorkflowService_GetReplicationMessages_Result) MethodName() string {
-	return "GetReplicationMessages"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Reply for this struct.
-func (v *WorkflowService_GetReplicationMessages_Result) EnvelopeType() wire.EnvelopeType {
 	return wire.Reply
 }
 
@@ -12474,645 +11188,6 @@ func (v *WorkflowService_QueryWorkflow_Result) MethodName() string {
 //
 // This will always be Reply for this struct.
 func (v *WorkflowService_QueryWorkflow_Result) EnvelopeType() wire.EnvelopeType {
-	return wire.Reply
-}
-
-// WorkflowService_ReapplyEvents_Args represents the arguments for the WorkflowService.ReapplyEvents function.
-//
-// The arguments for ReapplyEvents are sent and received over the wire as this struct.
-type WorkflowService_ReapplyEvents_Args struct {
-	ReapplyEventsRequest *shared.ReapplyEventsRequest `json:"reapplyEventsRequest,omitempty"`
-}
-
-// ToWire translates a WorkflowService_ReapplyEvents_Args struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_ReapplyEvents_Args) ToWire() (wire.Value, error) {
-	var (
-		fields [1]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.ReapplyEventsRequest != nil {
-		w, err = v.ReapplyEventsRequest.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-func _ReapplyEventsRequest_Read(w wire.Value) (*shared.ReapplyEventsRequest, error) {
-	var v shared.ReapplyEventsRequest
-	err := v.FromWire(w)
-	return &v, err
-}
-
-// FromWire deserializes a WorkflowService_ReapplyEvents_Args struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_ReapplyEvents_Args struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_ReapplyEvents_Args
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_ReapplyEvents_Args) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.ReapplyEventsRequest, err = _ReapplyEventsRequest_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_ReapplyEvents_Args
-// struct.
-func (v *WorkflowService_ReapplyEvents_Args) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [1]string
-	i := 0
-	if v.ReapplyEventsRequest != nil {
-		fields[i] = fmt.Sprintf("ReapplyEventsRequest: %v", v.ReapplyEventsRequest)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_ReapplyEvents_Args{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_ReapplyEvents_Args match the
-// provided WorkflowService_ReapplyEvents_Args.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_ReapplyEvents_Args) Equals(rhs *WorkflowService_ReapplyEvents_Args) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.ReapplyEventsRequest == nil && rhs.ReapplyEventsRequest == nil) || (v.ReapplyEventsRequest != nil && rhs.ReapplyEventsRequest != nil && v.ReapplyEventsRequest.Equals(rhs.ReapplyEventsRequest))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_ReapplyEvents_Args.
-func (v *WorkflowService_ReapplyEvents_Args) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.ReapplyEventsRequest != nil {
-		err = multierr.Append(err, enc.AddObject("reapplyEventsRequest", v.ReapplyEventsRequest))
-	}
-	return err
-}
-
-// GetReapplyEventsRequest returns the value of ReapplyEventsRequest if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Args) GetReapplyEventsRequest() (o *shared.ReapplyEventsRequest) {
-	if v != nil && v.ReapplyEventsRequest != nil {
-		return v.ReapplyEventsRequest
-	}
-
-	return
-}
-
-// IsSetReapplyEventsRequest returns true if ReapplyEventsRequest is not nil.
-func (v *WorkflowService_ReapplyEvents_Args) IsSetReapplyEventsRequest() bool {
-	return v != nil && v.ReapplyEventsRequest != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the arguments.
-//
-// This will always be "ReapplyEvents" for this struct.
-func (v *WorkflowService_ReapplyEvents_Args) MethodName() string {
-	return "ReapplyEvents"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Call for this struct.
-func (v *WorkflowService_ReapplyEvents_Args) EnvelopeType() wire.EnvelopeType {
-	return wire.Call
-}
-
-// WorkflowService_ReapplyEvents_Helper provides functions that aid in handling the
-// parameters and return values of the WorkflowService.ReapplyEvents
-// function.
-var WorkflowService_ReapplyEvents_Helper = struct {
-	// Args accepts the parameters of ReapplyEvents in-order and returns
-	// the arguments struct for the function.
-	Args func(
-		reapplyEventsRequest *shared.ReapplyEventsRequest,
-	) *WorkflowService_ReapplyEvents_Args
-
-	// IsException returns true if the given error can be thrown
-	// by ReapplyEvents.
-	//
-	// An error can be thrown by ReapplyEvents only if the
-	// corresponding exception type was mentioned in the 'throws'
-	// section for it in the Thrift file.
-	IsException func(error) bool
-
-	// WrapResponse returns the result struct for ReapplyEvents
-	// given the error returned by it. The provided error may
-	// be nil if ReapplyEvents did not fail.
-	//
-	// This allows mapping errors returned by ReapplyEvents into a
-	// serializable result struct. WrapResponse returns a
-	// non-nil error if the provided error cannot be thrown by
-	// ReapplyEvents
-	//
-	//   err := ReapplyEvents(args)
-	//   result, err := WorkflowService_ReapplyEvents_Helper.WrapResponse(err)
-	//   if err != nil {
-	//     return fmt.Errorf("unexpected error from ReapplyEvents: %v", err)
-	//   }
-	//   serialize(result)
-	WrapResponse func(error) (*WorkflowService_ReapplyEvents_Result, error)
-
-	// UnwrapResponse takes the result struct for ReapplyEvents
-	// and returns the erorr returned by it (if any).
-	//
-	// The error is non-nil only if ReapplyEvents threw an
-	// exception.
-	//
-	//   result := deserialize(bytes)
-	//   err := WorkflowService_ReapplyEvents_Helper.UnwrapResponse(result)
-	UnwrapResponse func(*WorkflowService_ReapplyEvents_Result) error
-}{}
-
-func init() {
-	WorkflowService_ReapplyEvents_Helper.Args = func(
-		reapplyEventsRequest *shared.ReapplyEventsRequest,
-	) *WorkflowService_ReapplyEvents_Args {
-		return &WorkflowService_ReapplyEvents_Args{
-			ReapplyEventsRequest: reapplyEventsRequest,
-		}
-	}
-
-	WorkflowService_ReapplyEvents_Helper.IsException = func(err error) bool {
-		switch err.(type) {
-		case *shared.BadRequestError:
-			return true
-		case *shared.DomainNotActiveError:
-			return true
-		case *shared.LimitExceededError:
-			return true
-		case *shared.ServiceBusyError:
-			return true
-		case *shared.EntityNotExistsError:
-			return true
-		default:
-			return false
-		}
-	}
-
-	WorkflowService_ReapplyEvents_Helper.WrapResponse = func(err error) (*WorkflowService_ReapplyEvents_Result, error) {
-		if err == nil {
-			return &WorkflowService_ReapplyEvents_Result{}, nil
-		}
-
-		switch e := err.(type) {
-		case *shared.BadRequestError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_ReapplyEvents_Result.BadRequestError")
-			}
-			return &WorkflowService_ReapplyEvents_Result{BadRequestError: e}, nil
-		case *shared.DomainNotActiveError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_ReapplyEvents_Result.DomainNotActiveError")
-			}
-			return &WorkflowService_ReapplyEvents_Result{DomainNotActiveError: e}, nil
-		case *shared.LimitExceededError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_ReapplyEvents_Result.LimitExceededError")
-			}
-			return &WorkflowService_ReapplyEvents_Result{LimitExceededError: e}, nil
-		case *shared.ServiceBusyError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_ReapplyEvents_Result.ServiceBusyError")
-			}
-			return &WorkflowService_ReapplyEvents_Result{ServiceBusyError: e}, nil
-		case *shared.EntityNotExistsError:
-			if e == nil {
-				return nil, errors.New("WrapResponse received non-nil error type with nil value for WorkflowService_ReapplyEvents_Result.EntityNotExistError")
-			}
-			return &WorkflowService_ReapplyEvents_Result{EntityNotExistError: e}, nil
-		}
-
-		return nil, err
-	}
-	WorkflowService_ReapplyEvents_Helper.UnwrapResponse = func(result *WorkflowService_ReapplyEvents_Result) (err error) {
-		if result.BadRequestError != nil {
-			err = result.BadRequestError
-			return
-		}
-		if result.DomainNotActiveError != nil {
-			err = result.DomainNotActiveError
-			return
-		}
-		if result.LimitExceededError != nil {
-			err = result.LimitExceededError
-			return
-		}
-		if result.ServiceBusyError != nil {
-			err = result.ServiceBusyError
-			return
-		}
-		if result.EntityNotExistError != nil {
-			err = result.EntityNotExistError
-			return
-		}
-		return
-	}
-
-}
-
-// WorkflowService_ReapplyEvents_Result represents the result of a WorkflowService.ReapplyEvents function call.
-//
-// The result of a ReapplyEvents execution is sent and received over the wire as this struct.
-type WorkflowService_ReapplyEvents_Result struct {
-	BadRequestError      *shared.BadRequestError      `json:"badRequestError,omitempty"`
-	DomainNotActiveError *shared.DomainNotActiveError `json:"domainNotActiveError,omitempty"`
-	LimitExceededError   *shared.LimitExceededError   `json:"limitExceededError,omitempty"`
-	ServiceBusyError     *shared.ServiceBusyError     `json:"serviceBusyError,omitempty"`
-	EntityNotExistError  *shared.EntityNotExistsError `json:"entityNotExistError,omitempty"`
-}
-
-// ToWire translates a WorkflowService_ReapplyEvents_Result struct into a Thrift-level intermediate
-// representation. This intermediate representation may be serialized
-// into bytes using a ThriftRW protocol implementation.
-//
-// An error is returned if the struct or any of its fields failed to
-// validate.
-//
-//   x, err := v.ToWire()
-//   if err != nil {
-//     return err
-//   }
-//
-//   if err := binaryProtocol.Encode(x, writer); err != nil {
-//     return err
-//   }
-func (v *WorkflowService_ReapplyEvents_Result) ToWire() (wire.Value, error) {
-	var (
-		fields [5]wire.Field
-		i      int = 0
-		w      wire.Value
-		err    error
-	)
-
-	if v.BadRequestError != nil {
-		w, err = v.BadRequestError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 1, Value: w}
-		i++
-	}
-	if v.DomainNotActiveError != nil {
-		w, err = v.DomainNotActiveError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 3, Value: w}
-		i++
-	}
-	if v.LimitExceededError != nil {
-		w, err = v.LimitExceededError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 4, Value: w}
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		w, err = v.ServiceBusyError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 5, Value: w}
-		i++
-	}
-	if v.EntityNotExistError != nil {
-		w, err = v.EntityNotExistError.ToWire()
-		if err != nil {
-			return w, err
-		}
-		fields[i] = wire.Field{ID: 6, Value: w}
-		i++
-	}
-
-	if i > 1 {
-		return wire.Value{}, fmt.Errorf("WorkflowService_ReapplyEvents_Result should have at most one field: got %v fields", i)
-	}
-
-	return wire.NewValueStruct(wire.Struct{Fields: fields[:i]}), nil
-}
-
-// FromWire deserializes a WorkflowService_ReapplyEvents_Result struct from its Thrift-level
-// representation. The Thrift-level representation may be obtained
-// from a ThriftRW protocol implementation.
-//
-// An error is returned if we were unable to build a WorkflowService_ReapplyEvents_Result struct
-// from the provided intermediate representation.
-//
-//   x, err := binaryProtocol.Decode(reader, wire.TStruct)
-//   if err != nil {
-//     return nil, err
-//   }
-//
-//   var v WorkflowService_ReapplyEvents_Result
-//   if err := v.FromWire(x); err != nil {
-//     return nil, err
-//   }
-//   return &v, nil
-func (v *WorkflowService_ReapplyEvents_Result) FromWire(w wire.Value) error {
-	var err error
-
-	for _, field := range w.GetStruct().Fields {
-		switch field.ID {
-		case 1:
-			if field.Value.Type() == wire.TStruct {
-				v.BadRequestError, err = _BadRequestError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 3:
-			if field.Value.Type() == wire.TStruct {
-				v.DomainNotActiveError, err = _DomainNotActiveError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 4:
-			if field.Value.Type() == wire.TStruct {
-				v.LimitExceededError, err = _LimitExceededError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 5:
-			if field.Value.Type() == wire.TStruct {
-				v.ServiceBusyError, err = _ServiceBusyError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		case 6:
-			if field.Value.Type() == wire.TStruct {
-				v.EntityNotExistError, err = _EntityNotExistsError_Read(field.Value)
-				if err != nil {
-					return err
-				}
-
-			}
-		}
-	}
-
-	count := 0
-	if v.BadRequestError != nil {
-		count++
-	}
-	if v.DomainNotActiveError != nil {
-		count++
-	}
-	if v.LimitExceededError != nil {
-		count++
-	}
-	if v.ServiceBusyError != nil {
-		count++
-	}
-	if v.EntityNotExistError != nil {
-		count++
-	}
-	if count > 1 {
-		return fmt.Errorf("WorkflowService_ReapplyEvents_Result should have at most one field: got %v fields", count)
-	}
-
-	return nil
-}
-
-// String returns a readable string representation of a WorkflowService_ReapplyEvents_Result
-// struct.
-func (v *WorkflowService_ReapplyEvents_Result) String() string {
-	if v == nil {
-		return "<nil>"
-	}
-
-	var fields [5]string
-	i := 0
-	if v.BadRequestError != nil {
-		fields[i] = fmt.Sprintf("BadRequestError: %v", v.BadRequestError)
-		i++
-	}
-	if v.DomainNotActiveError != nil {
-		fields[i] = fmt.Sprintf("DomainNotActiveError: %v", v.DomainNotActiveError)
-		i++
-	}
-	if v.LimitExceededError != nil {
-		fields[i] = fmt.Sprintf("LimitExceededError: %v", v.LimitExceededError)
-		i++
-	}
-	if v.ServiceBusyError != nil {
-		fields[i] = fmt.Sprintf("ServiceBusyError: %v", v.ServiceBusyError)
-		i++
-	}
-	if v.EntityNotExistError != nil {
-		fields[i] = fmt.Sprintf("EntityNotExistError: %v", v.EntityNotExistError)
-		i++
-	}
-
-	return fmt.Sprintf("WorkflowService_ReapplyEvents_Result{%v}", strings.Join(fields[:i], ", "))
-}
-
-// Equals returns true if all the fields of this WorkflowService_ReapplyEvents_Result match the
-// provided WorkflowService_ReapplyEvents_Result.
-//
-// This function performs a deep comparison.
-func (v *WorkflowService_ReapplyEvents_Result) Equals(rhs *WorkflowService_ReapplyEvents_Result) bool {
-	if v == nil {
-		return rhs == nil
-	} else if rhs == nil {
-		return false
-	}
-	if !((v.BadRequestError == nil && rhs.BadRequestError == nil) || (v.BadRequestError != nil && rhs.BadRequestError != nil && v.BadRequestError.Equals(rhs.BadRequestError))) {
-		return false
-	}
-	if !((v.DomainNotActiveError == nil && rhs.DomainNotActiveError == nil) || (v.DomainNotActiveError != nil && rhs.DomainNotActiveError != nil && v.DomainNotActiveError.Equals(rhs.DomainNotActiveError))) {
-		return false
-	}
-	if !((v.LimitExceededError == nil && rhs.LimitExceededError == nil) || (v.LimitExceededError != nil && rhs.LimitExceededError != nil && v.LimitExceededError.Equals(rhs.LimitExceededError))) {
-		return false
-	}
-	if !((v.ServiceBusyError == nil && rhs.ServiceBusyError == nil) || (v.ServiceBusyError != nil && rhs.ServiceBusyError != nil && v.ServiceBusyError.Equals(rhs.ServiceBusyError))) {
-		return false
-	}
-	if !((v.EntityNotExistError == nil && rhs.EntityNotExistError == nil) || (v.EntityNotExistError != nil && rhs.EntityNotExistError != nil && v.EntityNotExistError.Equals(rhs.EntityNotExistError))) {
-		return false
-	}
-
-	return true
-}
-
-// MarshalLogObject implements zapcore.ObjectMarshaler, enabling
-// fast logging of WorkflowService_ReapplyEvents_Result.
-func (v *WorkflowService_ReapplyEvents_Result) MarshalLogObject(enc zapcore.ObjectEncoder) (err error) {
-	if v == nil {
-		return nil
-	}
-	if v.BadRequestError != nil {
-		err = multierr.Append(err, enc.AddObject("badRequestError", v.BadRequestError))
-	}
-	if v.DomainNotActiveError != nil {
-		err = multierr.Append(err, enc.AddObject("domainNotActiveError", v.DomainNotActiveError))
-	}
-	if v.LimitExceededError != nil {
-		err = multierr.Append(err, enc.AddObject("limitExceededError", v.LimitExceededError))
-	}
-	if v.ServiceBusyError != nil {
-		err = multierr.Append(err, enc.AddObject("serviceBusyError", v.ServiceBusyError))
-	}
-	if v.EntityNotExistError != nil {
-		err = multierr.Append(err, enc.AddObject("entityNotExistError", v.EntityNotExistError))
-	}
-	return err
-}
-
-// GetBadRequestError returns the value of BadRequestError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Result) GetBadRequestError() (o *shared.BadRequestError) {
-	if v != nil && v.BadRequestError != nil {
-		return v.BadRequestError
-	}
-
-	return
-}
-
-// IsSetBadRequestError returns true if BadRequestError is not nil.
-func (v *WorkflowService_ReapplyEvents_Result) IsSetBadRequestError() bool {
-	return v != nil && v.BadRequestError != nil
-}
-
-// GetDomainNotActiveError returns the value of DomainNotActiveError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Result) GetDomainNotActiveError() (o *shared.DomainNotActiveError) {
-	if v != nil && v.DomainNotActiveError != nil {
-		return v.DomainNotActiveError
-	}
-
-	return
-}
-
-// IsSetDomainNotActiveError returns true if DomainNotActiveError is not nil.
-func (v *WorkflowService_ReapplyEvents_Result) IsSetDomainNotActiveError() bool {
-	return v != nil && v.DomainNotActiveError != nil
-}
-
-// GetLimitExceededError returns the value of LimitExceededError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Result) GetLimitExceededError() (o *shared.LimitExceededError) {
-	if v != nil && v.LimitExceededError != nil {
-		return v.LimitExceededError
-	}
-
-	return
-}
-
-// IsSetLimitExceededError returns true if LimitExceededError is not nil.
-func (v *WorkflowService_ReapplyEvents_Result) IsSetLimitExceededError() bool {
-	return v != nil && v.LimitExceededError != nil
-}
-
-// GetServiceBusyError returns the value of ServiceBusyError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Result) GetServiceBusyError() (o *shared.ServiceBusyError) {
-	if v != nil && v.ServiceBusyError != nil {
-		return v.ServiceBusyError
-	}
-
-	return
-}
-
-// IsSetServiceBusyError returns true if ServiceBusyError is not nil.
-func (v *WorkflowService_ReapplyEvents_Result) IsSetServiceBusyError() bool {
-	return v != nil && v.ServiceBusyError != nil
-}
-
-// GetEntityNotExistError returns the value of EntityNotExistError if it is set or its
-// zero value if it is unset.
-func (v *WorkflowService_ReapplyEvents_Result) GetEntityNotExistError() (o *shared.EntityNotExistsError) {
-	if v != nil && v.EntityNotExistError != nil {
-		return v.EntityNotExistError
-	}
-
-	return
-}
-
-// IsSetEntityNotExistError returns true if EntityNotExistError is not nil.
-func (v *WorkflowService_ReapplyEvents_Result) IsSetEntityNotExistError() bool {
-	return v != nil && v.EntityNotExistError != nil
-}
-
-// MethodName returns the name of the Thrift function as specified in
-// the IDL, for which this struct represent the result.
-//
-// This will always be "ReapplyEvents" for this struct.
-func (v *WorkflowService_ReapplyEvents_Result) MethodName() string {
-	return "ReapplyEvents"
-}
-
-// EnvelopeType returns the kind of value inside this struct.
-//
-// This will always be Reply for this struct.
-func (v *WorkflowService_ReapplyEvents_Result) EnvelopeType() wire.EnvelopeType {
 	return wire.Reply
 }
