+++ .gen/go/cadence/workflowserviceserver/server.go
@@ -28,7 +28,6 @@ package workflowserviceserver
 import (
 	context "context"
 	cadence "github.com/uber/cadence/.gen/go/cadence"
-	replicator "github.com/uber/cadence/.gen/go/replicator"
 	shared "github.com/uber/cadence/.gen/go/shared"
 	wire "go.uber.org/thriftrw/wire"
 	transport "go.uber.org/yarpc/api/transport"
@@ -65,16 +64,6 @@ type Interface interface {
 	GetClusterInfo(
 		ctx context.Context,
 	) (*shared.ClusterInfo, error)
-
-	GetDomainReplicationMessages(
-		ctx context.Context,
-		Request *replicator.GetDomainReplicationMessagesRequest,
-	) (*replicator.GetDomainReplicationMessagesResponse, error)
-
-	GetReplicationMessages(
-		ctx context.Context,
-		Request *replicator.GetReplicationMessagesRequest,
-	) (*replicator.GetReplicationMessagesResponse, error)
 
 	GetSearchAttributes(
 		ctx context.Context,
@@ -129,11 +118,6 @@ type Interface interface {
 		ctx context.Context,
 		QueryRequest *shared.QueryWorkflowRequest,
 	) (*shared.QueryWorkflowResponse, error)
-
-	ReapplyEvents(
-		ctx context.Context,
-		ReapplyEventsRequest *shared.ReapplyEventsRequest,
-	) error
 
 	RecordActivityTaskHeartbeat(
 		ctx context.Context,
@@ -319,28 +303,6 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 			},
 
 			thrift.Method{
-				Name: "GetDomainReplicationMessages",
-				HandlerSpec: thrift.HandlerSpec{
-
-					Type:  transport.Unary,
-					Unary: thrift.UnaryHandler(h.GetDomainReplicationMessages),
-				},
-				Signature:    "GetDomainReplicationMessages(Request *replicator.GetDomainReplicationMessagesRequest) (*replicator.GetDomainReplicationMessagesResponse)",
-				ThriftModule: cadence.ThriftModule,
-			},
-
-			thrift.Method{
-				Name: "GetReplicationMessages",
-				HandlerSpec: thrift.HandlerSpec{
-
-					Type:  transport.Unary,
-					Unary: thrift.UnaryHandler(h.GetReplicationMessages),
-				},
-				Signature:    "GetReplicationMessages(Request *replicator.GetReplicationMessagesRequest) (*replicator.GetReplicationMessagesResponse)",
-				ThriftModule: cadence.ThriftModule,
-			},
-
-			thrift.Method{
 				Name: "GetSearchAttributes",
 				HandlerSpec: thrift.HandlerSpec{
 
@@ -458,17 +420,6 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 					Unary: thrift.UnaryHandler(h.QueryWorkflow),
 				},
 				Signature:    "QueryWorkflow(QueryRequest *shared.QueryWorkflowRequest) (*shared.QueryWorkflowResponse)",
-				ThriftModule: cadence.ThriftModule,
-			},
-
-			thrift.Method{
-				Name: "ReapplyEvents",
-				HandlerSpec: thrift.HandlerSpec{
-
-					Type:  transport.Unary,
-					Unary: thrift.UnaryHandler(h.ReapplyEvents),
-				},
-				Signature:    "ReapplyEvents(ReapplyEventsRequest *shared.ReapplyEventsRequest)",
 				ThriftModule: cadence.ThriftModule,
 			},
 
@@ -705,7 +656,7 @@ func New(impl Interface, opts ...thrift.RegisterOption) []transport.Procedure {
 		},
 	}
 
-	procedures := make([]transport.Procedure, 0, 41)
+	procedures := make([]transport.Procedure, 0, 38)
 	procedures = append(procedures, thrift.BuildProcedures(service, opts...)...)
 	return procedures
 }
@@ -817,44 +768,6 @@ func (h handler) GetClusterInfo(ctx context.Context, body wire.Value) (thrift.Re
 
 	hadError := err != nil
 	result, err := cadence.WorkflowService_GetClusterInfo_Helper.WrapResponse(success, err)
-
-	var response thrift.Response
-	if err == nil {
-		response.IsApplicationError = hadError
-		response.Body = result
-	}
-	return response, err
-}
-
-func (h handler) GetDomainReplicationMessages(ctx context.Context, body wire.Value) (thrift.Response, error) {
-	var args cadence.WorkflowService_GetDomainReplicationMessages_Args
-	if err := args.FromWire(body); err != nil {
-		return thrift.Response{}, err
-	}
-
-	success, err := h.impl.GetDomainReplicationMessages(ctx, args.Request)
-
-	hadError := err != nil
-	result, err := cadence.WorkflowService_GetDomainReplicationMessages_Helper.WrapResponse(success, err)
-
-	var response thrift.Response
-	if err == nil {
-		response.IsApplicationError = hadError
-		response.Body = result
-	}
-	return response, err
-}
-
-func (h handler) GetReplicationMessages(ctx context.Context, body wire.Value) (thrift.Response, error) {
-	var args cadence.WorkflowService_GetReplicationMessages_Args
-	if err := args.FromWire(body); err != nil {
-		return thrift.Response{}, err
-	}
-
-	success, err := h.impl.GetReplicationMessages(ctx, args.Request)
-
-	hadError := err != nil
-	result, err := cadence.WorkflowService_GetReplicationMessages_Helper.WrapResponse(success, err)
 
 	var response thrift.Response
 	if err == nil {
@@ -1064,25 +977,6 @@ func (h handler) QueryWorkflow(ctx context.Context, body wire.Value) (thrift.Res
 
 	hadError := err != nil
 	result, err := cadence.WorkflowService_QueryWorkflow_Helper.WrapResponse(success, err)
-
-	var response thrift.Response
-	if err == nil {
-		response.IsApplicationError = hadError
-		response.Body = result
-	}
-	return response, err
-}
-
-func (h handler) ReapplyEvents(ctx context.Context, body wire.Value) (thrift.Response, error) {
-	var args cadence.WorkflowService_ReapplyEvents_Args
-	if err := args.FromWire(body); err != nil {
-		return thrift.Response{}, err
-	}
-
-	err := h.impl.ReapplyEvents(ctx, args.ReapplyEventsRequest)
-
-	hadError := err != nil
-	result, err := cadence.WorkflowService_ReapplyEvents_Helper.WrapResponse(err)
 
 	var response thrift.Response
 	if err == nil {
