+++ .gen/go/cadence/workflowservicetest/client.go
@@ -29,7 +29,6 @@ import (
 	context "context"
 	gomock "github.com/golang/mock/gomock"
 	workflowserviceclient "github.com/uber/cadence/.gen/go/cadence/workflowserviceclient"
-	replicator "github.com/uber/cadence/.gen/go/replicator"
 	shared "github.com/uber/cadence/.gen/go/shared"
 	yarpc "go.uber.org/yarpc"
 )
@@ -257,72 +256,6 @@ func (mr *_MockClientRecorder) GetClusterInfo(
 ) *gomock.Call {
 	args := append([]interface{}{ctx}, opts...)
 	return mr.mock.ctrl.RecordCall(mr.mock, "GetClusterInfo", args...)
-}
-
-// GetDomainReplicationMessages responds to a GetDomainReplicationMessages call based on the mock expectations. This
-// call will fail if the mock does not expect this call. Use EXPECT to expect
-// a call to this function.
-//
-// 	client.EXPECT().GetDomainReplicationMessages(gomock.Any(), ...).Return(...)
-// 	... := client.GetDomainReplicationMessages(...)
-func (m *MockClient) GetDomainReplicationMessages(
-	ctx context.Context,
-	_Request *replicator.GetDomainReplicationMessagesRequest,
-	opts ...yarpc.CallOption,
-) (success *replicator.GetDomainReplicationMessagesResponse, err error) {
-
-	args := []interface{}{ctx, _Request}
-	for _, o := range opts {
-		args = append(args, o)
-	}
-	i := 0
-	ret := m.ctrl.Call(m, "GetDomainReplicationMessages", args...)
-	success, _ = ret[i].(*replicator.GetDomainReplicationMessagesResponse)
-	i++
-	err, _ = ret[i].(error)
-	return
-}
-
-func (mr *_MockClientRecorder) GetDomainReplicationMessages(
-	ctx interface{},
-	_Request interface{},
-	opts ...interface{},
-) *gomock.Call {
-	args := append([]interface{}{ctx, _Request}, opts...)
-	return mr.mock.ctrl.RecordCall(mr.mock, "GetDomainReplicationMessages", args...)
-}
-
-// GetReplicationMessages responds to a GetReplicationMessages call based on the mock expectations. This
-// call will fail if the mock does not expect this call. Use EXPECT to expect
-// a call to this function.
-//
-// 	client.EXPECT().GetReplicationMessages(gomock.Any(), ...).Return(...)
-// 	... := client.GetReplicationMessages(...)
-func (m *MockClient) GetReplicationMessages(
-	ctx context.Context,
-	_Request *replicator.GetReplicationMessagesRequest,
-	opts ...yarpc.CallOption,
-) (success *replicator.GetReplicationMessagesResponse, err error) {
-
-	args := []interface{}{ctx, _Request}
-	for _, o := range opts {
-		args = append(args, o)
-	}
-	i := 0
-	ret := m.ctrl.Call(m, "GetReplicationMessages", args...)
-	success, _ = ret[i].(*replicator.GetReplicationMessagesResponse)
-	i++
-	err, _ = ret[i].(error)
-	return
-}
-
-func (mr *_MockClientRecorder) GetReplicationMessages(
-	ctx interface{},
-	_Request interface{},
-	opts ...interface{},
-) *gomock.Call {
-	args := append([]interface{}{ctx, _Request}, opts...)
-	return mr.mock.ctrl.RecordCall(mr.mock, "GetReplicationMessages", args...)
 }
 
 // GetSearchAttributes responds to a GetSearchAttributes call based on the mock expectations. This
@@ -684,37 +617,6 @@ func (mr *_MockClientRecorder) QueryWorkflow(
 ) *gomock.Call {
 	args := append([]interface{}{ctx, _QueryRequest}, opts...)
 	return mr.mock.ctrl.RecordCall(mr.mock, "QueryWorkflow", args...)
-}
-
-// ReapplyEvents responds to a ReapplyEvents call based on the mock expectations. This
-// call will fail if the mock does not expect this call. Use EXPECT to expect
-// a call to this function.
-//
-// 	client.EXPECT().ReapplyEvents(gomock.Any(), ...).Return(...)
-// 	... := client.ReapplyEvents(...)
-func (m *MockClient) ReapplyEvents(
-	ctx context.Context,
-	_ReapplyEventsRequest *shared.ReapplyEventsRequest,
-	opts ...yarpc.CallOption,
-) (err error) {
-
-	args := []interface{}{ctx, _ReapplyEventsRequest}
-	for _, o := range opts {
-		args = append(args, o)
-	}
-	i := 0
-	ret := m.ctrl.Call(m, "ReapplyEvents", args...)
-	err, _ = ret[i].(error)
-	return
-}
-
-func (mr *_MockClientRecorder) ReapplyEvents(
-	ctx interface{},
-	_ReapplyEventsRequest interface{},
-	opts ...interface{},
-) *gomock.Call {
-	args := append([]interface{}{ctx, _ReapplyEventsRequest}, opts...)
-	return mr.mock.ctrl.RecordCall(mr.mock, "ReapplyEvents", args...)
 }
 
 // RecordActivityTaskHeartbeat responds to a RecordActivityTaskHeartbeat call based on the mock expectations. This
