+++ client/frontend/retryableClient.go
@@ -25,7 +25,6 @@ import (
 
 	"go.uber.org/yarpc"
 
-	"github.com/uber/cadence/.gen/go/replicator"
 	"github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/common/backoff"
 )
@@ -568,48 +567,6 @@ func (c *retryableClient) UpdateDomain(
 	}
 	err := backoff.Retry(op, c.policy, c.isRetryable)
 	return resp, err
-}
-
-func (c *retryableClient) GetReplicationMessages(
-	ctx context.Context,
-	request *replicator.GetReplicationMessagesRequest,
-	opts ...yarpc.CallOption,
-) (*replicator.GetReplicationMessagesResponse, error) {
-	var resp *replicator.GetReplicationMessagesResponse
-	op := func() error {
-		var err error
-		resp, err = c.client.GetReplicationMessages(ctx, request, opts...)
-		return err
-	}
-	err := backoff.Retry(op, c.policy, c.isRetryable)
-	return resp, err
-}
-
-func (c *retryableClient) GetDomainReplicationMessages(
-	ctx context.Context,
-	request *replicator.GetDomainReplicationMessagesRequest,
-	opts ...yarpc.CallOption,
-) (*replicator.GetDomainReplicationMessagesResponse, error) {
-	var resp *replicator.GetDomainReplicationMessagesResponse
-	op := func() error {
-		var err error
-		resp, err = c.client.GetDomainReplicationMessages(ctx, request, opts...)
-		return err
-	}
-	err := backoff.Retry(op, c.policy, c.isRetryable)
-	return resp, err
-}
-
-func (c *retryableClient) ReapplyEvents(
-	ctx context.Context,
-	request *shared.ReapplyEventsRequest,
-	opts ...yarpc.CallOption,
-) error {
-
-	op := func() error {
-		return c.client.ReapplyEvents(ctx, request, opts...)
-	}
-	return backoff.Retry(op, c.policy, c.isRetryable)
 }
 
 func (c *retryableClient) GetClusterInfo(
