+++ idl/github.com/uber/cadence/cadence.thrift
@@ -554,34 +554,6 @@ service WorkflowService {
       6: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,
     )
 
-  replicator.GetReplicationMessagesResponse GetReplicationMessages(1: replicator.GetReplicationMessagesRequest request)
-    throws (
-      1: shared.BadRequestError badRequestError,
-      3: shared.LimitExceededError limitExceededError,
-      4: shared.ServiceBusyError serviceBusyError,
-      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,
-    )
-
-  replicator.GetDomainReplicationMessagesResponse GetDomainReplicationMessages(1: replicator.GetDomainReplicationMessagesRequest request)
-    throws (
-      1: shared.BadRequestError badRequestError,
-      3: shared.LimitExceededError limitExceededError,
-      4: shared.ServiceBusyError serviceBusyError,
-      5: shared.ClientVersionNotSupportedError clientVersionNotSupportedError,
-    )
-
-  /**
-  * ReapplyEvents applies stale events to the current workflow and current run
-  **/
-  void ReapplyEvents(1: shared.ReapplyEventsRequest reapplyEventsRequest)
-    throws (
-      1: shared.BadRequestError badRequestError,
-      3: shared.DomainNotActiveError domainNotActiveError,
-      4: shared.LimitExceededError limitExceededError,
-      5: shared.ServiceBusyError serviceBusyError,
-      6: shared.EntityNotExistsError entityNotExistError,
-    )
-
   /**
   * GetClusterInfo returns information about cadence cluster
   **/
